Win32: warn if the console font doesn't support Unicode
Unicode console output won't display correctly with default settings
because the default console font ("Terminal") only supports the system's
OEM charset. Unfortunately, this is a user specific setting, so it cannot
be easily fixed by e.g. some registry tricks in the setup program.

This change prints a warning on exit if console output contained non-ascii
characters and the console font is supposedly not a TrueType font (which
usually have decent Unicode support).

Signed-off-by: Karsten Blees <blees@dcon.de>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Stepan Kasal <kasal@ucw.cz>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e0b3e774612bb17695eb285774e0f1833f6e4901 [formerly e0b3e774612bb17695eb285774e0f1833f6e4901 [formerly 1edeb9abf5828e317999b4ebe8b7472c494341f2]]
Former-commit-id: 6801c9151d2fa419d004485e3da66f1d2a95c35f
Former-commit-id: 9ca70ecb000bf9ae0624123780b231f70fef8d83