git grep -O -i: if the pager is 'less', pass the '-I' option
When <command> happens to be the magic string "less", today

	git grep -O<command> -e<pattern>

helpfully passes +/<pattern> to less so you can navigate through
the results within a file using the n and shift+n keystrokes.

Alas, that doesn't do the right thing for a case-insensitive match,
i.e.

	git grep -i -O<command> -e<pattern>

For that case we should pass --IGNORE-CASE to "less" so that n and
shift+n can move between results ignoring case in the pattern.

The original patch came from msysgit and used "-i", but that was not
due to lack of support for "-I" but it merely overlooked that it
ought to work even when the pattern contains capital letters.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Stepan Kasal <kasal@ucw.cz>
Helped-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: cc5a2fd2fe3172bafe3529b226a298a825f514b0 [formerly 1adf0a47c188892e2aa641a2ebb3c86b6c8a8ad0] [formerly 1adf0a47c188892e2aa641a2ebb3c86b6c8a8ad0 [formerly f7febbea07cf6d65f85b0cdb6684d2e3666c557e]]
Former-commit-id: aa94bf1d6b92816c0dfd1a143b2f6aeb32fb22bc
Former-commit-id: d9964e3891d520237f98f87047c71e23feb66d02