bash prompt: use bash builtins to find out rebase state
During an ongoing interactive rebase __git_ps1() finds out the name of
the rebased branch, the total number of patches and the number of the
current patch by executing a '$(cat .git/rebase-merge/<FILE>)' command
substitution for each.  That is not quite the most efficient way to
read single line single word files, because it imposes the overhead of
fork()ing a subshell and fork()+exec()ing 'cat' several times.

Use the 'read' bash builtin instead to avoid those overheads.

Signed-off-by: SZEDER Gábor <szeder@ira.uka.de>

Former-commit-id: 74f0e366902a32188487d7143cb161fcd8ba3da7 [formerly 74f0e366902a32188487d7143cb161fcd8ba3da7 [formerly b91b935f04e8dcb1cc9f247627fbd0346ce949f4]]
Former-commit-id: 89f89612c066b4c249840d0bc408c0d970a00a44
Former-commit-id: 814c65531c87a2bca3f8f48992f779035a7ab2e7