bash prompt: use bash builtins to check stash state
When the environment variable $GIT_PS1_SHOWSTASHSTATE is set
__git_ps1() checks the presence of stashes by running 'git rev-parse
--verify refs/stash'.  This command not only checks that the
'refs/stash' ref exists but also, well, verifies that it's a valid
ref.

However, we don't need to be that thorough for the bash prompt.  We
can omit that verification and only check whether 'refs/stash' exists
or not.  Since 'git pack-refs' never packs 'refs/stash', it's a matter
of checking the existence of a ref file.  Perform this check using
only bash builtins to spare the overhead of fork()+exec()ing a git
process.

Also run 'git pack-refs --all' in the corresponding test to document
that the prompt script depends on 'git pack-refs' not packing
'refs/stash' and to catch possible breakages should this behavior ever
change.

Signed-off-by: SZEDER Gábor <szeder@ira.uka.de>

Former-commit-id: 4c8e98cf9cec7c4b48629e8244009ac5f3116edb [formerly 4c8e98cf9cec7c4b48629e8244009ac5f3116edb [formerly dd0b72cbd9e64c782a31c6acfca2ba9cf2ffb266]]
Former-commit-id: 826746fc0db6e7d04a7146aa9738c050094d2064
Former-commit-id: fe5913e9e233dd7c93539639c078ca598235f091