specifying ranges: we did not mean to make ".." an empty set
Either end of revision range operator can be omitted to default to HEAD,
as in "origin.." (what did I do since I forked) or "..origin" (what did
they do since I forked).  But the current parser interprets ".."  as an
empty range "HEAD..HEAD", and worse yet, because ".." does exist on the
filesystem, we get this annoying output:

  $ cd Documentation/howto
  $ git log .. ;# give me recent commits that touch Documentation/ area.
  fatal: ambiguous argument '..': both revision and filename
  Use '--' to separate filenames from revisions

Surely we could say "git log ../" or even "git log -- .." to disambiguate,
but we shouldn't have to.

Helped-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 63c1ef49efc3478bb0b0082362a57c0bb808f5f8 [formerly 63c1ef49efc3478bb0b0082362a57c0bb808f5f8 [formerly 003c84f6d2b9e9c4d5bbf5262cae994bac7190cb]]
Former-commit-id: 262e63fd426e1029e9479b2b244f765c13abd9b9
Former-commit-id: be913f6fe085deaab3d0b97294260505e898b552