gitk: Teach gitk to respect log.showroot
In early days, all projects managed by git (except for git itself) had the
product of a fairly mature development history in their first commit, and
it was deemed unnecessary clutter to show additions of these thousands of
paths as a patch.

"git log" learned to show the patch for the initial commit without requiring
--root command line option at 0f03ca9 (config option log.showroot to show
the diff of root commits, 2006-11-23).

Teach gitk to respect log.showroot.

[paulus@samba.org: Cleaned up the Tcl a bit, use --bool on the
 git config call]

Signed-off-by: Marcus Karlsson <mk@acc.umu.se>
Signed-off-by: Paul Mackerras <paulus@samba.org>

Former-commit-id: ed033ee4af6d798132bdff1e23a038e04e593ed8 [formerly 88b46cb788ee422d260374e63eb1e946807593bf] [formerly 88b46cb788ee422d260374e63eb1e946807593bf [formerly b2b76d10696d945bf19318831b64d009d119e051]]
Former-commit-id: ac3d3be39e61d613b5c13097caf9af17884d242e
Former-commit-id: 15f7dde43af0bf9fa2f23bcef03fb44346a24d2b