gitk: Speed up resolution of short SHA1 ids
On large repositories such as the Linux kernel, it can take quite a
noticeable time (several seconds) for gitk to resolve short SHA1 IDs
to their long form.  This speeds up the process by maintaining lists
of IDs indexed by the first 4 characters of the SHA1 ID, speeding up
the search by a factor of 65536 on large repositories.

Signed-off-by: Paul Mackerras <paulus@samba.org>

Former-commit-id: 7c8f91d5954c07a46e39c15386755319b6702420 [formerly 2a84f445aa83a84b43a7d72774af7a8a8fd1b851] [formerly 2a84f445aa83a84b43a7d72774af7a8a8fd1b851 [formerly 22387f2395217beaa7228cb02966cf7b61b0c346]]
Former-commit-id: 60f585abfe61567784643c4c099785023e441a86
Former-commit-id: 7738ed65e5f0004450e5220c5fe97346520b4514