gitk: Add menu items for comparing a commit with the marked commit
Sometimes one wants to see the different between two commits that are
a long distance apart in the graph display.  This is difficult to do
with the "Diff this -> selected" and "Diff selected -> this" menu
items because the need to maintain the selection means that one can't
use the find facilities or the reference list window to navigate from
one to the other.

This provides an alternative using the mark.  Having found one commit,
one marks it with the "Mark this commit" menu item, then navigates to
the other commit and uses the new "Diff this -> marked commit" and/or
"Diff marked commit -> this" menu items.

Signed-off-by: Paul Mackerras <paulus@samba.org>

Former-commit-id: c4aefb0e5d5127e0937473c829dcc4abebe8820c [formerly c4aefb0e5d5127e0937473c829dcc4abebe8820c [formerly 6febdede5acdbab192d955d4e0cd810f3fbee27f]]
Former-commit-id: 7cf9f7d44d58f55fa6a444fd92eba03df5137b98
Former-commit-id: 78e46add5d7af5252b0f5eb7fa4edb64cebe7ee0