configure: be more idiomatic
Lots of code in Git's configure.ac doesn't follow the typical formatting,
idioms and best practices for Autoconf input files.  Improve the situation.

There are probably many more similar improvements to be done, but trying
to clump all of them in a single change would make it unreviewable, so we
content ourselves with a partial improvement.

This change is just cosmetic, and should cause no semantic change.

The most relevant of the changes introduced by this patch are:

 - Do not add trailing '\' characters for line continuation where they
   are not truly needed.

 - In several (but not all) macro calls, properly quote the arguments.

 - Few cosmetic changes in spacing and comments.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 83b3b72284cb0c9b5369f6ec21e2983ff72cac0d [formerly 3078b81010c2ad13931b27b72011356f74611a6f] [formerly 3078b81010c2ad13931b27b72011356f74611a6f [formerly 5b2d131419817c5f5996da0e284e906400c083ef]]
Former-commit-id: 77d9419eb9045f1fcd5074cf744382c5c98be526
Former-commit-id: 0c9427f0771452b8ea1f0077e746bbddddb0b64d