git-svn: use POSIX::sigprocmask to block signals
In order to maintain consistency of the database mapping svn revision
numbers to git commit ids, rev_map_set() defers signal processing until
it's finished with an append transaction.[*]

The conventional way to achieve this is through sigprocmask(), which is
available in perl in the standard POSIX module.

This is implemented by this patch.  One important consequence of it is
that the signal handlers won't be unconditionally set to SIG_DFL anymore
upon the first invocation of rev_map_set() as they used to.  As a
result, the signals ignored by git-svn parent will remain ignored;
otherwise the behavior remains the same.

This patch paves the way to ignoring SIGPIPE throughout git-svn which
will be done in the followup patch.

[*] Deferring signals is not enough to ensure the database consistency:
the program may die on SIGKILL or power loss, run out of disk space,
etc.  However that's a separate issue that this patch doesn't address.

Signed-off-by: Roman Kagan <rkagan@mail.ru>
Acked-by: Eric Wong <normalperson@yhbt.net>

Former-commit-id: cf63c572d5b4f988ba011a9e98bbd5b2ef9f3bb2 [formerly cf63c572d5b4f988ba011a9e98bbd5b2ef9f3bb2 [formerly 037a98cd3f3efe85c6f56c4338002e4b2c7afa09]]
Former-commit-id: 2dbbcf6c852f6c5424d0f5698b28d1339e238202
Former-commit-id: c1be067c38f58a5f8dcc0e8821bafb27b895d4a6