git-svn: ignore SIGPIPE
In HTTP with keep-alive it's not uncommon for the client to notice that
the server decided to stop maintaining the current connection only when
sending a new request.  This naturally results in -EPIPE and possibly
SIGPIPE.

The subversion library itself makes no provision for SIGPIPE.  Some
combinations of the underlying libraries do (typically SIG_IGN-ing it),
some don't.

Presumably for that reason all subversion commands set SIGPIPE to
SIG_IGN early in their main()-s.

So should we.

This, together with the previous patch, fixes the notorious "git-svn
died of signal 13" problem (see e.g.
http://thread.gmane.org/gmane.comp.version-control.git/134936).

Signed-off-by: Roman Kagan <rkagan@mail.ru>
Acked-by: Eric Wong <normalperson@yhbt.net>

Former-commit-id: 44491b529ab6fba973d9b7f87a08f6a0d406c857 [formerly 44491b529ab6fba973d9b7f87a08f6a0d406c857 [formerly 6ade9bdadaf61565ee4e2ab47f66baaf41a20ecf]]
Former-commit-id: 3bbcc64acaa1bacbf214f918fd8ab9a63c751583
Former-commit-id: a6907de0104ba8fd635dcaac211f00d9b57f10f6