cache.h: hide on-disk index details
The on-disk format of the index file is a detail whose implementation is
neatly encapsulated in read-cache.c; there is no need to expose it to the
general public that include the cache.h header file.

Also add a prominent mark to read-cache.c to delineate the parts that deal
with the index file I/O routines from the remainder of the file.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 99c2ee3d0e78e656281f55d7ab1e282aead912df [formerly 99c2ee3d0e78e656281f55d7ab1e282aead912df [formerly db3b313c84522d89ac358187a7f6928dfcfb1e7d]]
Former-commit-id: e767dbe4cbce0c55192f9ac44bb502dd88a9b5eb
Former-commit-id: db02688992d725821866ff40fc11f4dbe65ca2fe