read-cache.c: allow unaligned mapping of the index file
Both the on-disk format v2 and v3 pads the "name" field to the multiple of
eight to make sure that various quantities in network long/short type can
be accessed with ntohl/ntohs without having to worry about alignment, but
this forces us to waste disk I/O bandwidth.

Introduce ntoh_s()/ntoh_l() macros that the callers can use as if they were
the regular ntohs()/ntohl() on a field that may not be aligned correctly.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7bf48a65b754c283e5cf5bf2602150c18089effd [formerly 7bf48a65b754c283e5cf5bf2602150c18089effd [formerly d60c49c2d7f683cd24a738533846cd3e34f2073e]]
Former-commit-id: 0b7f39ccfd16f153964c3538e09960f606a5a400
Former-commit-id: 2834f579dea638f5a013aafb2ff394c57c159593