read-cache.c: read prefix-compressed names in index on-disk version v4
Because the entries are sorted by path, adjacent entries in the index tend
to share the leading components of them, and it makes sense to only store
the differences in later entries.  In the v4 on-disk format of the index,
each on-disk cache entry stores the number of bytes to be stripped from
the end of the previous name, and the bytes to append to the result, to
come up with its name.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 28cd2d5c8e884aefe065ac53411e9633e36ba128 [formerly c6431a41d4eb9a5937b00d6266254837ffde0191] [formerly c6431a41d4eb9a5937b00d6266254837ffde0191 [formerly 6c9cd161d9d1bee349b1389d661282c92da098bc]]
Former-commit-id: aa33184b03a1edeca5b5eb90081621a45d837e35
Former-commit-id: 0a0be7f893057ab46b8d2610e70620595c82a9c9