gc: do not explode objects which will be immediately pruned
When we pack everything into one big pack with "git repack
-Ad", any unreferenced objects in to-be-deleted packs are
exploded into loose objects, with the intent that they will
be examined and possibly cleaned up by the next run of "git
prune".

Since the exploded objects will receive the mtime of the
pack from which they come, if the source pack is old, those
loose objects will end up pruned immediately. In that case,
it is much more efficient to skip the exploding step
entirely for these objects.

This patch teaches pack-objects to receive the expiration
information and avoid writing these objects out. It also
teaches "git gc" to pass the value of gc.pruneexpire to
repack (which in turn learns to pass it along to
pack-objects) so that this optimization happens
automatically during "git gc" and "git gc --auto".

Signed-off-by: Jeff King <peff@peff.net>
Acked-by: Nicolas Pitre <nico@fluxnic.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: dd8513ab2e7728ddab5cf19ed9adcd5663fdf354 [formerly 4f570b302ab758d240689509952889ffe8cd870b] [formerly 4f570b302ab758d240689509952889ffe8cd870b [formerly 7e52f5660e542cf801e8fc6902a30cc572c13736]]
Former-commit-id: 50a20ffc8f5103863b513fbdf71fb2f564777ae5
Former-commit-id: 7b393985bc68aed82404a7cfdacdd6d0adb3d5d2