fix http auth with multiple curl handles
HTTP authentication is currently handled by get_refs and fetch_ref, but
not by fetch_object, fetch_pack or fetch_alternates. In the
single-threaded case, this is not an issue, since get_refs is always
called first. It recognigzes the 401 and prompts the user for
credentials, which will then be used subsequently.

If the curl multi interface is used, however, only the multi handle used
by get_refs will have credentials configured. Requests made by other
handles fail with an authentication error.

Fix this by setting CURLOPT_USERPWD whenever a slot is requested.

Signed-off-by: Clemens Buchacher <drizzd@aon.at>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 075351ba4b57906ff588b05ecad62c791773c7da [formerly 075351ba4b57906ff588b05ecad62c791773c7da [formerly dfa1725a3ec57098637b698ffc2b2e2459acc518]]
Former-commit-id: 501e85c1e332369d9597392ff179c1e5c7f483c3
Former-commit-id: 62a899d05b8faa3cb9bfd3e0b5e70bd9618ba04c