git p4: import/export of labels to/from p4
The existing label import code looks at each commit being
imported, and then checks for labels at that commit. This
doesn't work in the real world though because it will drop
labels applied on changelists that have already been imported,
a common pattern.

This change adds a new --import-labels option. With this option,
at the end of the sync, git p4 gets sets of labels in p4 and git,
and then creates a git tag for each missing p4 label.

This means that tags created on older changelists are
still imported.

Tags that could not be imported are added to an ignore
list.

The same sets of git and p4 tags and labels can also be used to
derive a list of git tags to export to p4. This is enabled with
--export-labels in 'git p4 submit'.

Signed-off-by: Luke Diamand <luke@diamand.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 289133feaecb947f61a26c77cdd84e59eb23bb37 [formerly 289133feaecb947f61a26c77cdd84e59eb23bb37 [formerly 06804c76e8f7a61c7c9473decdf9b210d3238e9d]]
Former-commit-id: cf01eae1c14a19d24421703c58f01db8fe5daaf9
Former-commit-id: d7be566df3429611670bdfa6ccc509f5f39eca48