git-cherry-pick: add allow-empty option
git cherry-pick fails when picking a non-ff commit that is empty.  The advice
given with the failure is that a git-commit --allow-empty should be issued to
explicitly add the empty commit during the cherry pick.  This option allows a
user to specify before hand that they want to keep the empty commit.  This
eliminates the need to issue both a cherry pick and a commit operation.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 34b7006657abc05d0c99105cce7a0767c2048bf4 [formerly cd3983e016d21a5093f25007f38864268a0a696f] [formerly cd3983e016d21a5093f25007f38864268a0a696f [formerly df478b744cee9821eb2abbe2202d262b87a34289]]
Former-commit-id: 3d8fb270adc6b58a25f4aa1f5e74b804ef7a5e81
Former-commit-id: bf7f2c2ac610f15e7bdd53a0f5df7567703a1994