gitweb: Pass esc_html_hl_regions() options to esc_html()
With this change, esc_html_hl_regions() accepts options and passes them
down to esc_html().  This may be needed if a caller wants to pass
-nbsp=>1 to esc_html().

The idea and implementation example of this change was described in
337da8d2 (gitweb: Introduce esc_html_match_hl and esc_html_hl_regions,
2012-02-27).  While other suggestions may be more useful in some cases,
there is no need to implement them at the moment.  The
esc_html_hl_regions() interface may be changed later if it's needed.

[mk: extracted from larger patch and wrote commit message]

Signed-off-by: Jakub Narębski <jnareb@gmail.com>
Signed-off-by: Michał Kiedrowicz <michal.kiedrowicz@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 758b15a06e71e3e1d5deebe822d6a9d56241e9a5 [formerly ef405875001761a59a98757076eae076a19632a2] [formerly ef405875001761a59a98757076eae076a19632a2 [formerly 9768a9d884ee293548c0435df7645fd82d8d942e]]
Former-commit-id: ba40ece5fa7890e1609c93dbb36f72a84c47a947
Former-commit-id: d701e7f5d6481780a86b0411aec23e59457b274e