gitweb: Push formatting diff lines to print_diff_chunk()
Now lines are formatted closer to place where we actually use HTML
formatted output.

This means that we put raw lines in the @chunk accumulator, rather than
formatted lines.  Because we still need to know class (type) of line
when accumulating data to post-process and print, process_diff_line()
subroutine was retired and replaced by diff_line_class() used in
git_patchset_body() and new restructured format_diff_line() used in
print_diff_chunk().

As a side effect, we have to pass \%from and \%to down to callstack.

This is a preparation patch for diff refinement highlightning. It's not
meant to change gitweb output.

[jn: wrote commit message]

Signed-off-by: Michał Kiedrowicz <michal.kiedrowicz@gmail.com>
Acked-by: Jakub Narębski <jnareb@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2b57884977bee1e479f03c4a3fd351ff5ec13101 [formerly 2b57884977bee1e479f03c4a3fd351ff5ec13101 [formerly f4a81026508a48a129bd936aef2c75ac1f1cdd97]]
Former-commit-id: 060d66e1f7fd0b22b19273af5ec31947647a7870
Former-commit-id: b44f112ec6fca7d918a34342fdec018b8b3f4bbe