fast-import doc: cat-blob and ls responses need to be consumed quickly
If fast-import's command pipe and the frontend's cat-blob/ls response
pipe are both filled, there can be a deadlock.  Luckily all existing
frontends consume any pending cat-blob/ls responses completely before
writing the next command.

Document the requirements so future frontend authors and users can be
spared from the problem, too.  It is not always easy to catch that
kind of bug by testing.

To set the scene, add some words of explanation to help the novice
understand that "cat-blob" and "ls" output are meant for consumption
by the frontend.

Reported-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 708b55499f771f7400f092391342b92674fa30bf [formerly e0868854ff7ec05ed6aa50a7ce457f1e05922fe0] [formerly e0868854ff7ec05ed6aa50a7ce457f1e05922fe0 [formerly d57e490af31dff940052a5929fe05947c7ad03ad]]
Former-commit-id: 1e5af02dea83aa7b4583d528ebd30d60144cb272
Former-commit-id: d4514eaad111d5b2df28c5d5673a9319f758aa29