http: clean up leak in init_curl_http_auth
When we have a credential to give to curl, we must copy it
into a "user:pass" buffer and then hand the buffer to curl.
Old versions of curl did not copy the buffer, and we were
expected to keep it valid. Newer versions of curl will copy
the buffer.

Our solution was to use a strbuf and detach it, giving
ownership of the resulting buffer to curl. However, this
meant that we were leaking the buffer on newer versions of
curl, since curl was just copying it and throwing away the
string we passed. Furthermore, when we replaced a
credential (e.g., because our original one was rejected), we
were also leaking on both old and new versions of curl.

This got even worse in the last patch, which started
replacing the credential (and thus leaking) on every http
request.

Instead, let's use a static buffer to make the ownership
more clear and less leaky.  We already keep a static "struct
credential", so we are only handling a single credential at
a time, anyway.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ab8160669462f22cd60253847124429c0113a1a7 [formerly 009d93dda61e4aa309fd79942466e6eae53d62eb] [formerly 009d93dda61e4aa309fd79942466e6eae53d62eb [formerly aa0834a04e1d9d3ab81ecd4a4a138233e6e234ed]]
Former-commit-id: 73cb565f7e5b0e5141e3cb4529c5aeed98cf5943
Former-commit-id: 4066dafb76cd5dc7aab36e91d7941125c75e3213