http: use newer curl options for setting credentials
We give the username and password to curl by sticking them
in a buffer of the form "user:pass" and handing the result
to CURLOPT_USERPWD. Since curl 7.19.1, there is a split
mechanism, where you can specify each element individually.

This has the advantage that a username can contain a ":"
character. It also is less code for us, since we can hand
our strings over to curl directly. And since curl 7.17.0 and
higher promise to copy the strings for us, we we don't even
have to worry about memory ownership issues.

Unfortunately, we have to keep the ugly code for old curl
around, but as it is now nicely #if'd out, we can easily get
rid of it when we decide that 7.19.1 is "old enough".

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ea042bb8e84e2c7fbea67108112f35957c3f7f2c [formerly ea042bb8e84e2c7fbea67108112f35957c3f7f2c [formerly 6f4c347ca1d3102d77e2dd36b6bc8ab12de6045b]]
Former-commit-id: 7ef40c0c8092b5875673c75085f6754093f4fb26
Former-commit-id: fe6492a7d95cb8428f777f7b119027a5982dc5bc