Provide branch name in error message when using @{u}
When using @{u} or @{upstream} it is common to omit the branch name,
implying current branch. If the upstream is not configured, the error
message was "No upstream branch found for ''".

When resolving '@{u}', branch_get() is called, which almost always
returns a description of a branch. This allows us to use a branch name
in the error message, even if the user said something like '@{u}'.

The only case when branch_get() returns NULL is when HEAD points to so
something which is not a branch. Of course this also means that no
upstream is configured, but it is better to directly say that HEAD
does not point to a branch.

Signed-off-by: Zbigniew Jędrzejewski-Szmek <zbyszek@in.waw.pl>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c3ce31dd885186281b70cd7feba0121ff172f2ee [formerly 868ccabf756770aec2bb49c67f052652ec6c1a1a] [formerly 868ccabf756770aec2bb49c67f052652ec6c1a1a [formerly 9884e67f9d9b511592b981572cfbb30cc50aec61]]
Former-commit-id: 97fa3a4eef3cd720c156ab0811a090e688f0cea8
Former-commit-id: 8bfd88bb10260b5ceca217aa18cd0327723742af