Be more specific if upstream branch is not tracked
If the branch configured as upstream didn't have a local tracking
branch, git said "Upstream branch not found". We can be more helpful,
and separate the cases when upstream is not configured, and when it is
configured, but the upstream branch is not tracked in a local branch.

The following configuration leads to the second scenario:

    [remote "origin"]
    	    url = ...
            fetch = refs/heads/master
    [branch "master"]
            remote = origin
            merge = refs/heads/master

'git pull' will work on master, but master@{upstream} is not defined.

Signed-off-by: Zbigniew Jędrzejewski-Szmek <zbyszek@in.waw.pl>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c71511ad0e972622b0218bd057d3a47a1afb08ee [formerly c71511ad0e972622b0218bd057d3a47a1afb08ee [formerly 17c82211ec2d29a6e7221948cba8dae9074dc5a6]]
Former-commit-id: 3c49a6b93624077626263d6287562492fc4c46a6
Former-commit-id: e19403cdf87a4d5cb7abb66607d22d8c74f97f98