cherry-pick: do not expect file arguments
If a commit-ish passed to cherry-pick or revert happens to have a file
of the same name, git complains that the argument is ambiguous and
advises to use '--'. To make things worse, the '--' argument is removed
by parse_options, und so passing '--' has no effect.

Instead, always interpret cherry-pick/revert arguments as revisions.

Signed-off-by: Clemens Buchacher <drizzd@aon.at>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b4f475b2c46b9e8d74b923fd669c314b8933efee [formerly ff2af4bb8ae4788172c6a99c3e17a63c4a7ddc0b] [formerly ff2af4bb8ae4788172c6a99c3e17a63c4a7ddc0b [formerly 6d5b93f29f5362b9b84954d87684705227984f9e]]
Former-commit-id: 022df1cc8c51c6a7f7c2ea0f6c4a4ad5a8f680cb
Former-commit-id: a5dd9085788af0333ad2e087ccc6e9e1268e8144