argv-array: refactor empty_argv initialization
An empty argv-array is initialized to point to a static
empty NULL-terminated array.  The original implementation
separates the actual storage of the NULL-terminator from the
pointer to the list.  This makes the exposed type a "const
char **", which nicely matches the type stored by the
argv-array.

However, this indirection means that one cannot use
empty_argv to initialize a static variable, since it is
not a constant.

Instead, we can expose empty_argv directly, as an array of
pointers. The only place we use it is in the ARGV_ARRAY_INIT
initializer, and it decays to a pointer appropriately there.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 19db5932202bc0ec0806b2d87f066fc4f78bb67d [formerly 19db5932202bc0ec0806b2d87f066fc4f78bb67d [formerly fd93d2e60ea66fc3796904ff53ead3ef4755b137]]
Former-commit-id: 315523c148f537b9007b99e4bad7001b0dd63c83
Former-commit-id: 0e4006a980eb202f60709b4f453f83a288d2d1c2