gc: use argv-array for sub-commands
git-gc executes many sub-commands. The argument list for
some of these is constant, but for others we add more
arguments at runtime. The latter is implemented by allocating
a constant extra number of NULLs, and either using a custom
append function, or just referencing unused slots by number.

As of commit 7e52f56, which added two new arguments, it is
possible to exceed the constant number of slots for "repack"
by running "git gc --aggressive", causing "git gc" to die.

This patch converts all of the static argv lists to use
argv-array. In addition to fixing the overflow caused by
7e52f56, it has a few advantages:

  1. We can drop the custom append function (which,
     incidentally, had an off-by-one error exacerbating the
     static limit).

  2. We can drop the ugly magic numbers used when adding
     arguments to "prune".

  3. Adding further arguments will be easier; you can just
     add new "push" calls without worrying about increasing
     any static limits.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: fc71c72488cdd464c421b4437a367b9149489fda [formerly 8931e4c43aab2640f7ca9ba4e6c87a3162b5b85b] [formerly 8931e4c43aab2640f7ca9ba4e6c87a3162b5b85b [formerly 234587fc87b156dc20461fc61353beeb904b43bc]]
Former-commit-id: c29511e20c65ee9ec67352d5f6a35493b4158864
Former-commit-id: d4bff77c84b1f42f009c2ccde9b40538b16db7a7