t9300-fast-import: avoid 'exit' in test_expect_success snippets
Exiting from a for-loop early using '|| break' does not propagate the
failure code, and for this reason, the tests used just 'exit'. But this
ends the test script with 'FATAL: Unexpected exit code 1' in the case of
a failed test.

Fix this by moving the loop into a shell function, from which we can
simply return early.

While at it, modernize the style of the affected test cases.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 77aa3a3c417bdf25b80993b0e63e8d034e06b7e8 [formerly 77aa3a3c417bdf25b80993b0e63e8d034e06b7e8 [formerly 05880b02221dc8da136e9febfa7853ad822b1ee1]]
Former-commit-id: d73ac0bc53a34113a8ff1f365eaf5ad9df3f6329
Former-commit-id: 3b4131fe0671499dc914bd35c38fd9f58067051a