git-rebase: add keep_empty flag
Add a command line switch to git-rebase to allow a user the ability to specify
that they want to keep any commits in a series that are empty.

When git-rebase's type is am, then this option will automatically keep any
commit that has a tree object identical to its parent.

This patch changes the default behavior of interactive rebases as well.  With
this patch, git-rebase -i will produce a revision set passed to
git-revision-editor, in which empty commits are commented out.  Empty commits
may be kept manually by uncommenting them.  If the new --keep-empty option is
used in an interactive rebase the empty commits will automatically all be
uncommented in the editor.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 42a0691d4bf6d52a913571b5d77d2ce95f5567b0 [formerly 42a0691d4bf6d52a913571b5d77d2ce95f5567b0 [formerly 90e1818f9a06015159712e204dd90868e0a6c421]]
Former-commit-id: fda40d33b7c6315cc9adf4591a22cbf88a40bd03
Former-commit-id: eb0236f9353ecb33295dd5067d8d3ef4d9d2cc21