t5570: use explicit push refspec
The default mode for push without arguments will change. Some warnings
are about to be enabled for such use, which causes some t5570 tests to
fail because they do not expect this output.

Fix this by passing an explicit refspec to git push. To that end, change
the calling conventions of test_remote_error in order to accomodate
extra command arguments.

Signed-off-by: Clemens Buchacher <drizzd@aon.at>
Signed-off-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0baa60351a6400eea290382e7023c2e2ea78f16a [formerly 040fee617c96667b4cdde58fecb190637d962efd] [formerly 040fee617c96667b4cdde58fecb190637d962efd [formerly aecff47da67af69dae4e53749277ca2728e31e36]]
Former-commit-id: c03a1ad855c6b33eef333f77af2b9aeeb25a7b3b
Former-commit-id: 9cbf8dbacabdc17e8745ef79262cd791060cf085