get_ref_dir(): take the containing directory as argument
Previously, the "dir" argument to get_ref_dir() was a pointer to the
top-level ref_dir.  Change the function to expect a pointer to the
ref_dir corresponding to dirname.  This allows entries to be added
directly to dir, without having to recurse through the reference trie
each time (i.e., we can use add_entry_to_dir() instead of add_ref()).

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 56ca6c72f37db524ba2999dea7e84e4a952dfcb9 [formerly 56ca6c72f37db524ba2999dea7e84e4a952dfcb9 [formerly 9f2fb4a3737c86e245ce365f6ad8f901ad397d6f]]
Former-commit-id: 7275597fe26815e68ba45b0ad62f57b0788b3564
Former-commit-id: 5177e5d0644176748d9522f189148b2f28639bf2