help: replace underlining "help -a" headers using hyphens with a blank line
We used to underline a header text, like this:

    This is a header
    ----------------
    content...

But calculating text length so that the dashes align with the text
could get complicated because the text could be in any charset in
translated Git.

There is no point to use this pseudo underline; simply a blank
line would do and it even makes it easier to read:

    This is a header

    content...

While at it, give translators more context to translate, e.g.
e.g.  "git commands available..." instead of "%s available..."

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c3fdeaf7b0feb1d8914b6700dbc9ee2f207b062b [formerly c3fdeaf7b0feb1d8914b6700dbc9ee2f207b062b [formerly 4470ef94973cf40bb83864d480cee1e064053879]]
Former-commit-id: 773a983829fef951aa13a29cc52c57eddb926f0c
Former-commit-id: f6331d336bc95ddd8c3a287f9ffef49f973bd3a1