remove blank filename in error message
When write_loose_object() finds that it is unable to
create a temporary file, it complains, for instance:

    unable to create temporary sha1 filename : Too many open files

That extra space was supposed to be the name of the file,
and will be an empty string if the git_mkstemps_mode() fails.

The name of the temporary file is unimportant; delete it.

Signed-off-by: Pete Wyckoff <pw@padd.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c92cb724f9b839d857dc982ca478ff4ee2353f6b [formerly 2749255282cd20605939d856f128b15fc5ebdebf] [formerly 2749255282cd20605939d856f128b15fc5ebdebf [formerly 5eaeda70de1e7763d070e8ccfbcd8ca10baea800]]
Former-commit-id: 5ba891433d48fd270a5a466427acd7e77df6fa5e
Former-commit-id: 1b12343d459f8b845aaa9978e989af69e0288736