git p4: fix writable file after rename or copy
The way rename works is with a "p4 integrate", optionally
followed by a "p4 edit" if the change is not a 100% rename.
Contents are generated by applying a patch, not doing a file
system rename.  Copy is similar.

In this case, p4 does not fix the permissions back to read-only.
Make sure this happens by calling "p4 sync -f".

Signed-off-by: Pete Wyckoff <pw@padd.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 77e497edd9e009e0d05d685cad50510e4d5df117 [formerly 77e497edd9e009e0d05d685cad50510e4d5df117 [formerly b6ad6dcc3b0629d525abc9fe0882e1b0eb969e17]]
Former-commit-id: a37f29bff0ecbea06413dd3831945b43c5133231
Former-commit-id: d4a26244d647843fa2beea15f87e149cde0cff09