archive-zip: streaming for deflated files
After an entry has been streamed out, its CRC and sizes are written as
part of a data descriptor.

For simplicity, we make the buffer for the compressed chunks twice as
big as for the uncompressed ones, to be sure the result fit in even
if deflate makes them bigger.

t5000 verifies output. t1050 makes sure the command always respects
core.bigfilethreshold

Signed-off-by: Rene Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 27a7158a5ba5ed169ae6ba9c322efbd17352b146 [formerly 27a7158a5ba5ed169ae6ba9c322efbd17352b146 [formerly c743c21591f9433fe784ac38902872701ce2e850]]
Former-commit-id: 784ffcc790a494f4fea30ec1eb31c2d095fc2d2b
Former-commit-id: 6a15d9261a668980db3b473dc10fa30ea76eea12