checkout (detached): truncate list of orphaned commits at the new HEAD
When git checkout switches from a detached HEAD to any other commit, then
all orphaned commits were listed in a warning:

  Warning: you are leaving 2 commits behind...:

    a5e5396 another fixup
    6aa1af6 fixup foo

But if the new commit is actually one from this list (6aa1af6 in this
example), then the list in the warning can be truncated at the new HEAD,
because history beginning at HEAD is not "left behind". This makes it so.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1a82f9929c353dc281a57074d17c003aedf388f6 [formerly 11a216c3401bdc17dd0132aae80a4386f2213c75] [formerly 11a216c3401bdc17dd0132aae80a4386f2213c75 [formerly 5d8863954f077d2c262d5cc4fc669947ff23d6d5]]
Former-commit-id: 008821b55aa8eeeebc37b0e721085b0cd14b8038
Former-commit-id: b0152789e76598793d344d6db3eed329899333ba