status: refactor colopts handling
The current code reads the config and command-line options
into a separate "colopts" variable, and then copies the
contents of that variable into the "struct wt_status". We
can eliminate the extra variable and copy just write
straight into the wt_status struct.

This simplifies the "status" code a little bit.
Unfortunately, it makes the "commit" code one line more
complex; a side effect of the separate variable was that
"commit" did not copy the colopts variable, so any
column.status configuration had no effect.

The result still ends up cleaner, though. In the previous
version, it was unclear whether commit simply forgot to copy
the colopt variable, or whether it was intentional. Now it
explicitly turns off column options. Furthermore, if commit
later learns to respect column.status, this will make the
end result simpler. I punted on just adding that feature
now, because it was sufficiently non-obvious that it should
not go into a refactoring patch.

Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: 476724e0b7008830f1f82a6039fde999638ce4f6 [formerly 476724e0b7008830f1f82a6039fde999638ce4f6 [formerly 4d2292e9a90465cbfa23f1c170c11423d11830df]]
Former-commit-id: 5f10f4cb8bfe9d6eae66d8467d0572a2245d586d
Former-commit-id: 848bb6bfa972f0c9d9c5ae513d9e89d9888d3300