status: respect "-b" for porcelain format
There is no reason not to, as the user has to explicitly ask
for it, so we are not breaking compatibility by doing so. We
can do this simply by moving the "show_branch" flag into
the wt_status struct. As a bonus, this saves us from passing
it explicitly, simplifying the code.

Signed-off-by: Jeff King <peff@peff.net>

Former-commit-id: 9ea49e102cac5c09924c8cef022d4e3a2f6d59ea [formerly 9ea49e102cac5c09924c8cef022d4e3a2f6d59ea [formerly d4a6bf1fb64d904e210fbf7c5b330b06438a5bd5]]
Former-commit-id: 4e9946bd93a69899b4350b120fb95c106fbc36a7
Former-commit-id: 911c128bb7bdede0a1623ccb6053a3f787249fbb