apply: clear_image() clears things a bit more
The clear_image() function did not clear the line table in the image
structure; this does not matter for the current callers, as the function
is only called from the codepaths that deal with binary patches where the
line table is never populated, and the codepaths that do populate the line
table free it themselves.

But it will start to matter when we introduce a codepath to retry a failed
patch, so make sure it clears and frees everything.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 434019efb8a48aa922d700bdb7ff4332c5417b7c [formerly 70f4b7e9c81e9f1d5ba5c05276ff109d03919ca6] [formerly 70f4b7e9c81e9f1d5ba5c05276ff109d03919ca6 [formerly 798b9ce87bfcd18896f9b0e21918b7bf31109b32]]
Former-commit-id: 9dc3c623687d429db5c8aec10a2010b1da064dd4
Former-commit-id: dea0f550d8df16f8d151edfc7fcaa42b6903d368