apply: refactor read_file_or_gitlink()
Reading a blob out of the object store does not have to require that the
caller has a cache entry for it.

Create a read_blob_object() helper function that takes the object name and
mode, and use it to reimplement the original function as a thin wrapper to
it.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 88016970123a68ab08611383f2ffadd445b057f1 [formerly a147c29f4d49a2245b7518cd62e4f00f8a766e18] [formerly a147c29f4d49a2245b7518cd62e4f00f8a766e18 [formerly e42a96e7726365557b5865957927dc1ca272cad2]]
Former-commit-id: 72988cd6e7c7967259f13b707cb4b11cf1a71ae7
Former-commit-id: 46ad4c484aee6d2bb7b9570ea4ca1a514ad0037e