xdiff: remove unused functions
The functions xdl_cha_first(), xdl_cha_next() and xdl_atol() are not used
by us.  While removing them increases the difference to the upstream
version of libxdiff, it only adds a bit to the more than 600 differing
lines in xutils.c (mmfile_t management was simplified significantly when
the library was imported initially).  Besides, if upstream modifies these
functions in the future, we won't need to think about importing those
changes, so in that sense it makes tracking modifications easier.

Signed-off-by: Rene Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e1fdfe6692cb7d5953ac3647a91002e17cf3d8df [formerly e1fdfe6692cb7d5953ac3647a91002e17cf3d8df [formerly be89977543c7798411f336236a506bf07b0a3d71]]
Former-commit-id: d694f9beac3d8d959fce09710b336a8bfdfeeeee
Former-commit-id: 21f4da6b339eaea5730fac4da8daaf96e9c96be5