apply: plug the three-way merge logic in
When a patch does not apply to what we have, but we know the preimage the
patch was made against, we apply the patch to the preimage to compute what
the patch author wanted the result to look like, and attempt a three-way
merge between the result and our version, using the intended preimage as
the base version.

When we are applying the patch using the index, we would additionally need
to add the object names of these three blobs involved in the merge, which
is not yet done in this step, but we add a field to "struct patch" so that
later write-out step can use it.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 733588e385b0824c98ade8e8c04937c756ff5647 [formerly 733588e385b0824c98ade8e8c04937c756ff5647 [formerly 28ff051268463e93613998a017a10d3de45abbc0]]
Former-commit-id: dc491b9ba8452b63a0419a4822cf2450cf0f8158
Former-commit-id: ec903e0205675ae52574475597d41363bc85485e