apply: register conflicted stages to the index
Now we have all the necessary logic to fall back on three-way merge when
the patch does not cleanly apply, insert the conflicted entries to the
index as appropriate.  This obviously triggers only when the "--index"
option is used.

When we fall back to three-way merge and some of the merges fail, just
like the case where the "--reject" option was specified and we had to
write some "*.rej" files out for unapplicable patches, exit the command
with non-zero status without showing the diffstat and summary.  Otherwise
they would make the list of problematic paths scroll off the display.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: fa2f9107d9dbe1ed21554e53c2932d75f18c003d [formerly fa2f9107d9dbe1ed21554e53c2932d75f18c003d [formerly 4f4a6cb9882516d75f345772c9bf00b891753f10]]
Former-commit-id: 71717d0d6fcfc3979931a1ebaa9b0c6799104a6e
Former-commit-id: 02f66dc198c5b1bfbe43a1c48b5b29e4ce1f10b4