document submdule.$name.update=none option for gitmodules
This option was not yet described in the gitmodules documentation. We
only described it in the 'git submodule' command documentation but
gitmodules is the more natural place to look.

A short reference in the 'git submodule' documentation should be
sufficient since the details can now be found in the documentation to
gitmodules.

Signed-off-by: Heiko Voigt <hvoigt@hvoigt.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: db5742ad41dfe43f0c5304514670f82979874183 [formerly db5742ad41dfe43f0c5304514670f82979874183 [formerly e6a1c43aafa8b221e05e57132549dcb3cd7866cf]]
Former-commit-id: f18f2fc37b0758304467b0c815becd907be32d33
Former-commit-id: 3adf9f8a483ba830200d57038dbd9c4fec0bc634