dir: respect string length argument of read_directory_recursive()
A directory name is passed to read_directory_recursive() as a
length-limited string, through the parameters base and baselen.
Suprisingly, base must be a NUL-terminated string as well, as it is
passed to opendir(), ignoring baselen.

Fix this by postponing the call to opendir() until the length-limted
string is added to a strbuf, which provides a NUL in the right place.

Signed-off-by: Rene Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 26f5ac38e5a94cf4dda5058be61651cd8e1f371c [formerly d2664af91fca3f3711eefb358f24c729fe15bda1] [formerly d2664af91fca3f3711eefb358f24c729fe15bda1 [formerly 1528d247e53bcf33db5b004a25a7d9b8c75c90e4]]
Former-commit-id: 77849230f55a23b1aec312785ce0ca456d63ea75
Former-commit-id: e2b2226b449a231a375b1b12fe04fbbf12c25601