git-svn: support rebase --preserve-merges
When git svn rebase is performed after an unpushed merge, the
rebase operation follows both parents and replays both the user's
local commits and those from the merged branch. This is usually
not the intended behavior.
This patch adds support for the --preserve-merges/-p flag which
allows for a better workflow by re-applying merge commits as merges.

[ew: fixed a minor syntax error]

Signed-off-by: Avishay Lavie <avishay.lavie@gmail.com>
Signed-off-by: Eric Wong <normalperson@yhbt.net>

Former-commit-id: 129bc2cffd60e5a943b86a36176699dd347cb39b [formerly 129bc2cffd60e5a943b86a36176699dd347cb39b [formerly b64e1f58158d1d1a8eafabbbf002a1a3c1d72929]]
Former-commit-id: 88ddf6f4558ca7d7423f7043510f8b74b8765689
Former-commit-id: 537be515eca0db3f021550d36a6126031bedca0a