pack-objects: refactor write_object() into helper functions
The function first decides if we want to copy data taken from existing
pack verbatim or we want to encode the data ourselves for the packfile
we are creating and then carries out the decision.  Separate the latter
phase into two helper functions, one for the case the data is reused,
the other for the case the data is produced anew.

A little twist is that it can later turn out that we cannot reuse the
data after we initially decide to do so; in such a case, the "reuse"
helper makes a call to "generate" helper.  It is easier to follow than
the current fallback code that uses "goto" inside a single large
function.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d0d558ee37efef769a855e504de7f2a75eaec60f [formerly d0d558ee37efef769a855e504de7f2a75eaec60f [formerly c9018b0305a56436c85b292edbeacff04b0ebb5d]]
Former-commit-id: e13d7d10c248d6a663069e9d1a8fee920444d6fe
Former-commit-id: bb3264eab4c2143b326d3806620f1770bad4c4a7