git-sh-setup: define workaround wrappers before they are used
Recently we tweaked this scriptlet to let mingw port redefine "pwd" to
always return Windows-style path, but the code to do so came after the
first use of "pwd" to set up $GIT_DIR shell variable.

Move the block to define these workaround wrappers, so that everything
everything that executes when the scriptlet is dot-sourced uses the
replacements.

Noticed-by: Ramsay Jones
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7ce65ba6b5e7d438f8028e4226036e2631de9d22 [formerly 7ce65ba6b5e7d438f8028e4226036e2631de9d22 [formerly 10587ce6dfb479cf4930657e173b673031deaa06]]
Former-commit-id: 0cc87f57b1dad112110d1af51c73bdbc12c14215
Former-commit-id: c8683ade928bcaf0ca0ca2787ec98860a0201915