apply: refactor "previous patch" logic
The code to grab the result of application of a previous patch in the
input was mixed with error message generation for a case where a later
patch tries to modify contents of a path that has been removed.

The same code is duplicated elsewhere in the code.  Introduce a helper
to clarify what is going on.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6266c6f19e7b7837f203190236e1ecaebd5cc783 [formerly 6266c6f19e7b7837f203190236e1ecaebd5cc783 [formerly ccf998b297c6d04053eaa9d61e9cb4187bccd0a9]]
Former-commit-id: 46065b7318f957e5c637dc7566f44cb578d6d34d
Former-commit-id: 4e526f77ad81cee8556555764539f46e405d5eaf