git-svn: clarify the referent of dcommit's optional argument
The documentation of the dcommit subcommand is reworded to clarify that
the optional argument refers to a git branch, not an SVN branch.

The discussion of the optional argument is put into its own paragraph
as is the guidance about using 'dcommit' in preference to 'set-tree'.

The section on REBASE vs. PULL/MERGE is reworded to incorporate the
advice to prefer 'git rebase' previously in the description of 'dcommit'.

Signed-off-by: Jon Seymour <jon.seymour@gmail.com>
Acked-by: Eric Wong <normalperson@yhbt.net>

Former-commit-id: b436ee1e45a22a1045e19f57365c819d8b1e82ef [formerly b436ee1e45a22a1045e19f57365c819d8b1e82ef [formerly 2b26b65f9abc77c4af87626452005a73edda0c8f]]
Former-commit-id: fea8a88c3554650f41ee2b02f93afc0b4ba1cee3
Former-commit-id: 4d595f931a722ff4b4968e5472627bca30ad08af