completion: add support for backwards compatibility
Some people might be relying on _git and _gitk to define custom aliases,
unfortunately, commit 6b179ad (completion: add new __git_complete
helper) broke that support.

  "bash: [: 1: unary operator expected"

This can be easily fixed by using __git_complete, but it's not meant to
be public.

Although _git and _gitk are probably not meant to be public, it's easy
to keep having support for them by having a wrapper to the proper
new function that is fully functional.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 044335df650983c656e5a8b5d9a1d5a81fa57f1f [formerly 044335df650983c656e5a8b5d9a1d5a81fa57f1f [formerly b0a4b2d2570974f2050b5d43828dd8146505753e]]
Former-commit-id: 68d3563ff7dbf5f5ed882389b7b27e8e016544a1
Former-commit-id: 1cb0fc86407c5dc5bba576f1efd1469a78b6acf1