Consistently use "superproject" instead of "supermodule"
We fairly consistently say "superproject" and never "supermodule" these
days. But there are seven occurrences of "supermodule" left in the current
work tree. Three appear in Release Notes for 1.5.3 and 1.7.7, three in
test names and one in a C-code comment.

Replace all occurrences of "supermodule" outside of the Release Notes
(which shouldn't be changed after the fact) with "superproject" for
consistency.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a9c786556e8a8168c262b993758e5fc5cce95ba3 [formerly a9c786556e8a8168c262b993758e5fc5cce95ba3 [formerly cb8ad289c6b3f774e6e4431a08eb585eadbb8dad]]
Former-commit-id: d7c64bce78bc53d5d53105438429e05646eb6b94
Former-commit-id: bd51a4ec4e0cc040ee44a6e5e68c4a57a2118aa7