fetch-pack: sort the list of incoming refs
Having the list sorted means we can avoid some quadratic
algorithms when comparing lists.

These should typically be sorted already, but they do come
from the remote, so let's be extra careful. Our ref-sorting
implementation does a mergesort, so we do not have to care
about performance degrading in the common case that the list
is already sorted.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a9d5bc116f347fc384f3a487e82d38cb6ee07b44 [formerly a9d5bc116f347fc384f3a487e82d38cb6ee07b44 [formerly 9e8e704f0bc14da325aa63f639c6dc782c81e26f]]
Former-commit-id: 51416bc5ecb3d17853e1334b89ec1d0e730b4a3f
Former-commit-id: 7b5596ef56535cb93d639e64c484329cbd9dc4e8