fetch-pack: avoid quadratic loop in filter_refs
We have a list of refs that we want to compare against the
"match" array. The current code searches the match list
linearly, giving quadratic behavior over the number of refs
when you want to fetch all of them.

Instead, we can compare the lists as we go, giving us linear
behavior.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ac16b0c2f2da75352f5e29fd06a5ae2f5458a977 [formerly ac16b0c2f2da75352f5e29fd06a5ae2f5458a977 [formerly a0de28805dceaca86a6f7bedfd3e8c227b781d9d]]
Former-commit-id: 726dc0d720bd63ae041ef1cf64ef87d57fbd1c80
Former-commit-id: d835a8a6b620f0cfbceb5ff8c468e47690b5bb91