ident: split setup_ident into separate functions
This function sets up the default name, email, and date, and
is not publicly available. Let's split it into three public
functions so that callers can get just the parts they need.

While we're at it, let's change the interface to simple
accessors. The original function was called only by fmt_ident,
and contained logic for "if we already have some other
value, don't load the default" which properly belongs in
fmt_ident.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 97e9ac94cfc4e3724d1d9e3648cd0d464e26a4be [formerly 97e9ac94cfc4e3724d1d9e3648cd0d464e26a4be [formerly bcb2b0044b3baedf7857613ec069f300c364680c]]
Former-commit-id: fa7406893b381de49209d8a090741da9639fa076
Former-commit-id: 2d5bb07195e9ac22687e53bc8df50937acf5a397