fmt-merge-msg: don't use static buffer in record_person
The record_person function just parses out the "name" field
of the person line in a commit and adds it to a string_list.
The only reason we need an extra buffer is that the
string_list functions require a NUL-terminated string.

Instead of the static buffer, we can just allocate a
temporary NUL-terminated copy. In addition to removing a
useless limit, this removes the only user of MAX_GITNAME
outside of ident.c.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 48a42ff390b608a6319dbfdf46d509689da0bc51 [formerly 48a42ff390b608a6319dbfdf46d509689da0bc51 [formerly e21ab1340a1a22f017e9c3e4fc0deb9f4de36917]]
Former-commit-id: 0acbb54929b17b3643628bf6e5b7e13936bd8cc9
Former-commit-id: 92b0b01b7749eb476ff81bbbb354411547e41bed