ident: trim trailing newline from /etc/mailname
We use fgets to read the /etc/mailname file, which means we
will typically end up with an extra newline in our
git_default_email. Most of the time this doesn't matter, as
fmt_ident will skip it as cruft, but there is one code path
that accesses it directly (in http-push.c:lock_remote).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a0080a7551d8611968493ad0605c2ab4c6eaaad2 [formerly a0080a7551d8611968493ad0605c2ab4c6eaaad2 [formerly 132f4b6ccb470cb209167b7806c68805ba4dc600]]
Former-commit-id: 74c205551a3e83b23ce279bddd421476c68bf1fa
Former-commit-id: 17271bedafcfac934aab037dded2c2f752f2e020