format-patch: use default email for generating message ids
We try to generate a sane message id for cover letters and
threading by appending some changing bits to the front of
the user's email address. The current code parses the email
out of the results of git_committer_info, but we can do this
much more easily by just calling ident_default_email
ourselves.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a19741b0b57acc3af8c5daada431eb28459dc58c [formerly a19741b0b57acc3af8c5daada431eb28459dc58c [formerly 43ae9f47ab8a7762d914e91d6f57b79126986640]]
Former-commit-id: 3460815c54de66a124e38d56952a70d6f2933047
Former-commit-id: cce85080f42b070ed2130ea3cb3088c571ae946f