format-patch: refactor get_patch_filename
The get_patch_filename function expects a commit argument
and uses it to get the sanitized subject line when making a
patch filename. However, we also want to use this same
function for the cover letter, which does not have a commit
object. The current solution is to create a fake commit with
the subject "cover letter". Instead, let's make the
get_patch_filename interface more flexibile, and allow
passing a direct subject.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 14ec552435e1d4304a61938119962a86117a699b [formerly cb1dd5e8931a674a5138b276f78b309b5f1a3dcc] [formerly cb1dd5e8931a674a5138b276f78b309b5f1a3dcc [formerly a21c2f94fb6e74ebddb5c78cf6b5f68983646530]]
Former-commit-id: 9d15a0f9af5d80e2adae3c7d92aa5864129bcc56
Former-commit-id: d7fc6773aaf271ef7003fcb00f49dac202c077ff