osxkeychain: pull make config from top-level directory
The default compiler and cflags were mostly "works for me"
when I built the original version. We need to be much less
careful here than usual, because we know we are building
only on OS X.  But it's only polite to at least respect the
CFLAGS and CC definitions that the user may have provided
earlier.

While we're at it, let's update our definitions and rules to
be more like the top-level Makefile; default our CFLAGS to
include -O2, and make sure we use CFLAGS and LDFLAGS when
linking.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a41cd7f4fbd2ee874850e04e9eb690903a112b95 [formerly a41cd7f4fbd2ee874850e04e9eb690903a112b95 [formerly 17a9ac7d6b5f27eb5f8efee324e88db2b6ec9bd6]]
Former-commit-id: b0f6891f009ec7a3c131ad4e5ef0b34e57d6f993
Former-commit-id: e01d66e417bd29477b82d0d0d8c80c245cac6add