ident: reword empty ident error message
There's on point in printing the name, since it is by
definition the empty string if we have reached this code
path. Instead, let's be more clear that we are complaining
about the empty name, but still show the email address that
it is attached to (since that may provide some context to
the user).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 165e41ffa36db17816ea08b90260552b542ff340 [formerly 165e41ffa36db17816ea08b90260552b542ff340 [formerly b00f6cfcd7232d90c4625c42eb9694d4ed2dc615]]
Former-commit-id: 43102d6ede4e8fab927d12a40003433442d93c6e
Former-commit-id: c9228a3234baba90fe0dbd225fb5550cda6f59a1