ident: refactor NO_DATE flag in fmt_ident
As a short-hand, we extract this flag into the local
variable "name_addr_only". It's more accurate to simply
negate this and refer to it as "want_date", which will be
less confusing when we add more NO_* flags.

While we're touching this part of the code, let's move the
call to ident_default_date() only when we are actually going
to use it, not when we have NO_DATE set, or when we get a
date from the environment.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4ad2c820055819ff52b0a7eccd8015417d6858b7 [formerly 4ad2c820055819ff52b0a7eccd8015417d6858b7 [formerly 359b27add341878a13c6a8b85849b75b78246c7e]]
Former-commit-id: 2a4b6ba93bfb1c7cd2039602ac8cf0ae3213ab6d
Former-commit-id: beb3a851c0f9e083b188bbe345ff665063b539dd