ident: let callers omit name with fmt_indent
Most callers want to see all of "$name <$email> $date", but
a few want only limited parts, omitting the date, or even
the name. We already have IDENT_NO_DATE to handle the date
part, but there's not a good option for getting just the
email. Callers have to done one of:

  1. Call ident_default_email; this does not respect
     environment variables, nor does it promise to trim
     whitespace or other crud from the result.

  2. Call git_{committer,author}_info; this returns the name
     and email, leaving the caller to parse out the wanted
     bits.

This patch adds IDENT_NO_NAME; it stops short of adding
IDENT_NO_EMAIL, as no callers want it (nor are likely to),
and it complicates the error handling of the function.

When no name is requested, the angle brackets (<>) around
the email address are also omitted.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1f2ec02ab165feaff981d6357f0a17d79fc0fa21 [formerly 1f2ec02ab165feaff981d6357f0a17d79fc0fa21 [formerly c15e1987aec562107bd8e4a9fdeebf6d27d0e53a]]
Former-commit-id: 85f40bc7f1f868861ea30155cafa6efa65b39011
Former-commit-id: a0c54dbe160417d5d8dd5676a7c22cad91728ead