format-patch: use GIT_COMMITTER_EMAIL in message ids
Before commit 43ae9f4, we generated the tail of a message id
by calling git_committer_info and parsing the email out of
the result. 43ae9f4 changed to use ident_default_email
directly, so we didn't have to bother with parsing. As a
side effect, it meant we no longer used GIT_COMMITTER_EMAIL
at all.

In general, this is probably reasonable behavior. Either the
default email is sane on your system, or you are using
user.email to provide something sane. The exception is if
you rely on GIT_COMMITTER_EMAIL being set all the time to
override the bogus generated email.

This is unlikely to match anybody's real-life setup, but we
do use it in the test environment. And furthermore, it's
what we have always done, and the change in 43ae9f4 was
about cleaning up, not fixing any bug; we should be
conservative and keep the behavior identical.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 76ed628c37b053302648f7911218272957267684 [formerly 4b8f11b3f69ef397b91ac40bc75002a9b37fa307] [formerly 4b8f11b3f69ef397b91ac40bc75002a9b37fa307 [formerly c73f384f92dcb0d7111533ed21f6c2dc9fcc323b]]
Former-commit-id: 1708e4cff616d23c91901f66399beaead0e8014e
Former-commit-id: e8cdaddd835c6f3328115a23472610b51114d425