ident: rename IDENT_ERROR_ON_NO_NAME to IDENT_STRICT
Callers who ask for ERROR_ON_NO_NAME are not so much
concerned that the name will be blank (because, after all,
we will fall back to using the username), but rather it is a
check to make sure that low-quality identities do not end up
in things like commit messages or emails (whereas it is OK
for them to end up in things like reflogs).

When future commits add more quality checks on the identity,
each of these callers would want to use those checks, too.
Rather than modify each of them later to add a new flag,
let's refactor the flag.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4f8df23b20ebc6305bec41ef046501578cc73633 [formerly df4cfd0d9d309464badbca5202e52c9f074b06d3] [formerly df4cfd0d9d309464badbca5202e52c9f074b06d3 [formerly f9bc573fdaeaf8621008f3f49aaaa64869791691]]
Former-commit-id: e37cfeedf4d1a3b3911ebdc42a379958b0e98570
Former-commit-id: f8f8f881f70931f0b0d82f54c6fdd081eef6079b