fmt-merge-message: add empty line between tag and signature verification
When adding the information from a tag, put an empty line between the
message of the tag and the commented-out signature verification
information.

At least for the kernel workflow, I often end up re-formatting the message
that people send me in the tag data. In that situation, putting the tag
message and the tag signature verification back-to-back then means that
normal editor "reflow parapgraph" command will get confused and think that
the signature is a continuation of the last message paragraph.

So I always end up having to first add an empty line, and then go back and
reflow the last paragraph. Let's just do it in git directly.

The extra vertical space also makes the verification visually stand out
more from the user-supplied message, so it looks a bit more readable to me
too, but that may be just an odd personal preference.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c7f068a587c4cf45c350a62728df740977d0b6d7 [formerly c7f068a587c4cf45c350a62728df740977d0b6d7 [formerly a3347b988a338e95b7f3b11eda776ac0d7b84dd0]]
Former-commit-id: 0ba81a5e38e3417e1ebd0244ddb10f9e20993446
Former-commit-id: 459ea5dab295cecebf84009ee8a4f6350e3410e7