config doc: remove confusion about relative GIT_DIR from FILES section
From the FILES section of the git-config(1) manual:

	$GIT_DIR/config::
		Repository specific configuration file. (The filename is
		of course relative to the repository root, not the working
		directory.)

That's confusing because $GIT_DIR really is relative to the working
directory.

	$ GIT_DIR=.git GIT_EDITOR='pwd; echo editing'
	$ export GIT_DIR GIT_EDITOR
	$ git config --edit --local
	/home/jrn/src/git/Documentation
	editing .git/config

It turns out that the comment is a remnant from older days when the
heading said ".git/config" (which is indeed relative to the top of the
worktree).

It was only when the heading was changed to refer more precisely to
<git dir>/config (see v1.5.3.2~18, AsciiDoc tweak to avoid leading
dot, 2007-09-14) that the parenthesis stopped making sense.  Remove
it.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0d0f10ad1723d600ddcf34686dd22ae2d7e1be9e [formerly 0d0f10ad1723d600ddcf34686dd22ae2d7e1be9e [formerly ec84e069afd7845b1d6bb37d2aaefa913b0e0141]]
Former-commit-id: 474487769cfb5180bf0832e0235417fdec257bcd
Former-commit-id: d9246b857003a450ff1ff2ebafd10160e293794b