Reduce cost of deletion in levenstein distance (4 -> 3)
Before this patch, a character deletion has the same cost as 2 swaps, or
4 additions, so Git prefers suggesting a completely scrambled command
name to removing a character. For example, "git tags" suggests "stage",
but not "tag".

By setting the deletion cost to 3, we keep it higher than swaps or
additions, but prefer 1 deletion to 2 swaps. "git tags" now suggests
"tag" in addition to staged.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e5b4c82193512acbb2cdd142a570f32238bec64e [formerly e5b4c82193512acbb2cdd142a570f32238bec64e [formerly c41494f8c89c330cc9bd4629cdc3c428803e76bc]]
Former-commit-id: d1f1be5b8a4aa421628203243b840181154e0f4e
Former-commit-id: e3dd5f0b4ed2bda0bd9cd4bc92ca20d714e8917a