git-svn: rename SVN::Git::* packages to Git::SVN::*
Using names in the Git:: namespace means these cannot conflict with a
hypothetical binding teaching Subversion to interact with git
repositories.

Currently the packages are private to git-svn.perl so the choice of
name isn't likely to make much difference.  This change is mainly
meant as preparation for splitting out the packages in question as
modules on the public search path.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Eric Wong <normalperson@yhbt.net>

Former-commit-id: 69f2ea5355216feec9ad94182adf92789c942ad2 [formerly 69f2ea5355216feec9ad94182adf92789c942ad2 [formerly 72827aa468b467157643bb3689c8786e86e4a58c]]
Former-commit-id: 85ea2f27361557990a4dac3e5d081d2916e30938
Former-commit-id: 474de2423e2a509ecb2997f1d5b5d053f3673896