cherry-pick: regression fix for empty commits
The earlier "--keep-redundant-commit" series broke "cherry-pick"
that is given a commit whose change is already in the current
history. Such a cherry-pick would result in an empty change, and
should stop with an error, telling the user that conflict resolution
may have made the result empty (which is exactly what is happening),
but we silently dropped the change on the floor without any message
nor non-zero exit code.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0ad5103e011b39bc3f18f69ca353bd5990063528 [formerly 0ad5103e011b39bc3f18f69ca353bd5990063528 [formerly ac2b0e8fb77d02f480af3f45a44cb9ea611008a4]]
Former-commit-id: b3d55beda954084fc3a8bad08d5dbc6f31d04c66
Former-commit-id: e2496cc4b916fc40c6ac2011dc7c4fcd58956573