l10n: Set nplurals of zh_CN.po from 1 to 2
In most cases, plural-forms are unnecessary for Chinese. For example,
"apple" and "apples" are the same in Chinese, they are both translated
as "苹果". While there are exceptions, e.g., the plural form of "he",
"she" and "it" is "they" in English. In Chinese, "他(he)", "她(she)",
and "它(it)" have plural forms too, they are "他们", "她们", and "它们".

But what makes 'nplurals=1' hard to work right for Chinese is:

    #: bundle.c:192
    #, c-format
    msgid "The bundle requires this ref"
    msgid_plural "The bundle requires these %d refs"

Signed-off-by: Jiang Xin <worldhello.net@gmail.com>

Former-commit-id: 2aadce8d47bc1d1a01f78d4fd41eac3a08da9817 [formerly 2aadce8d47bc1d1a01f78d4fd41eac3a08da9817 [formerly 5ae481e0ad61f3cdf0b0f6bc3f57bae23574db16]]
Former-commit-id: 7b94e58c527901c18ca0b190ce7b0146e9ae97d8
Former-commit-id: 12ee441f9a5e6ca1285aa24d1c248c56ff440523