request-pull: really favor a matching tag
After tagging the tip of "dev" branch with a "for-linus" tag and
pushing both out, running

	$ git request-pull $url $last_release dev

would produce an output asking the 'dev' branch of $url to be
pulled, because that is what the user asked the message to say.

We already detect this situation locally and include the contents of
the tag in the output; if the $url has that tag, favor that tag
(i.e. "for-linus") in the generated message over the branch name the
user gave us (i.e. "dev") from the command line, to make the output
look more consistent.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: cfe7993c694011a6e6c109c2f1e4dac38914ca98 [formerly cfe7993c694011a6e6c109c2f1e4dac38914ca98 [formerly 682853e687ff0106bea25bdb5e9d60d9657cecb3]]
Former-commit-id: 5a8e15e28b2de8aac6a5b6b481bda45a3e03a3e3
Former-commit-id: 193de3f8da88bcd9abd194b9f5693aad7eecfe0b