pager: drop "wait for output to run less" hack
Commit 35ce862 (pager: Work around window resizing bug in
'less', 2007-01-24) causes git's pager sub-process to wait
to receive input after forking but before exec-ing the
pager. To handle this, run-command had to grow a "pre-exec
callback" feature. Unfortunately, this feature does not work
at all on Windows (where we do not fork), and interacts
poorly with run-command's parent notification system. Its
use should be discouraged.

The bug in less was fixed in version 406, which was released
in June 2007. It is probably safe at this point to remove
our workaround. That lets us rip out the preexec_cb feature
entirely.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 233dcd9e071ef8d3b4891f6e8a93016efa610636 [formerly 233dcd9e071ef8d3b4891f6e8a93016efa610636 [formerly e8320f350f523c3219ff8ec639663193941af57d]]
Former-commit-id: 2bb8bc237aab12cdd270da032d24bcd09e401a95
Former-commit-id: fcac26e4d8728afc69ebb6ebfe0bc6e996dca47d