submodule: fix sync handling of some relative superproject origin URLs
When the origin URL of the superproject is itself relative, git submodule sync
configures the remote.origin.url configuration property of the submodule
with a path that is relative to the work tree of the superproject
rather than the work tree of the submodule.

To fix this an 'up_path' that navigates from the work tree of the submodule
to the work tree of the superproject needs to be prepended to the URL
otherwise calculated.

Correct handling of superproject origin URLs like foo, ./foo and ./foo/bar is
left to a subsequent patch since an additional change is required to handle
these cases.

The documentation of resolve_relative_url() is expanded to give a more thorough
description of the function's objective.

Signed-off-by: Jon Seymour <jon.seymour@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 36f0bf7ed7160f256cf42455c47b94b4b65ca6ee [formerly 27ffb8144d580bc72eaefc68f966247d976f9b47] [formerly 27ffb8144d580bc72eaefc68f966247d976f9b47 [formerly 967b2c66738fe6b168ead0c3106ac47fb9ae22c7]]
Former-commit-id: 3701aa4b97d3fbd0aa42943629f047b031f63dd6
Former-commit-id: 5cd6eafd6829d92c2502757ba9c70cb73f21e168