submodule: fix handling of superproject origin URLs like foo, ./foo and ./foo/bar
Currently git submodule init and git submodule sync fail with an error
if the superproject origin URL is of the form foo but succeed if the
superproject origin URL is of the form ./foo or ./foo/bar or foo/bar.

This change makes handling of the foo case behave like the handling
of the ./foo case and also ensures that superfluous leading and
embedded ./'s are removed from the resulting derived URLs.

Signed-off-by: Jon Seymour <jon.seymour@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f8a94e7db7d4cf18b34227ab6e70bd86ce815bc7 [formerly f8a94e7db7d4cf18b34227ab6e70bd86ce815bc7 [formerly 758615e2517db8b9fda9218afded06a3e1b42e20]]
Former-commit-id: 342892a99fa997f956ec98838ad9778d3f5e5acc
Former-commit-id: bc4e8ad4ef65cbf2c8206273b3a132aded55dfe7