help: use HTML as the default help format on Windows
When 'git help $cmd' is run without a format option (e.g. -w), the
'man' format is always used. On some platforms, however, manual page
viewers are not often available.

Introduce DEFAULT_HELP_FORMAT make variable in order to allow the
default format configurable at compile time, and set it to HTML when
compiling on Windows (but not Cygwin).

Helped-by: Jeff King <peff@peff.net>
Signed-off-by: Vincent van Ravesteijn <vfr@lyx.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 9624a5e04d3a23e704cf9a9fe24fcd64e40d7fd6 [formerly 9624a5e04d3a23e704cf9a9fe24fcd64e40d7fd6 [formerly 1cc8af044cad37b5f7df85b177f6aa979aa3215a]]
Former-commit-id: b91a0eb656a5295b021d53b45b3138b97ef4766c
Former-commit-id: 43daa676111cdfa514076ab205f487cb9e7d5a4f