docs: fix cross-directory linkgit references
Most of our documentation is in a single directory, so using
linkgit:git-config[1] just generates a relative link in the
same directory. However, this is not the case with the API
documentation in technical/*, which need to refer to
git-config from the parent directory.

We can fix this by passing a special prefix attribute when building
in a subdirectory, and respecting that prefix in our linkgit
definitions.

We only have to modify the html linkgit definition.  For
manpages, we can ignore this for two reasons:

  1. we do not generate actual links to the file in
     manpages, but instead just give the name and section of
     the linked manpage

  2. we do not currently build manpages for subdirectories,
     only html

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 153c1324f42a0d2e979a909aa2c15b1ecfde1ebb [formerly 153c1324f42a0d2e979a909aa2c15b1ecfde1ebb [formerly fe77b416c774a469a585c4b58fcae5e7a5f3545d]]
Former-commit-id: f47c65ad79ef9e421273c03e8fe3f0c36a5d5def
Former-commit-id: dcc0746f9519c548fc151c5ca4fca582850aee47