apply: move "already exists" logic to check_to_create()
The check_to_create_blob() function used to check only the case
where we are applying to the working tree.  Rename the function to
check_to_create() and make it also responsible for checking the case
where we apply to the index.  Also make its caller responsible for
issuing an error message.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 124d0713920b3a2af35cd6a100f4cdb235a5d5ec [formerly 124d0713920b3a2af35cd6a100f4cdb235a5d5ec [formerly ec15be02679ed5a98e4a96b23b99dd1b90196fae]]
Former-commit-id: dae3b2a674148a24743c2ddca87a49ecf6f688ef
Former-commit-id: 07dfbd1be5a5681538dc46e27e2cb7de735540e1