apply: --3way with add/add conflict
When a patch wants to create a path, but we already have it in our
current state, pretend as if the patch and we independently added
the same path and cause add/add conflict, so that the user can
resolve it just like "git merge" in the same situation.

For that purpose, implement load_current() in terms of the
load_patch_target() helper introduced earlier to read the current
contents from the path given by patch->new_name (patch->old_name is
NULL for a creation patch).

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 14ee244727caa962b215882adaf4545537166780 [formerly 14ee244727caa962b215882adaf4545537166780 [formerly 099f3c421a446e2eef3a1bb6aea0e451711ddd88]]
Former-commit-id: 67bfdbeee419fbf5cb604e78aaf6c62b9b5cb07b
Former-commit-id: b6cfe50c71c8c1402096a391958876f4f44a734d