update-index: allow overwriting existing submodule index entries
In commit e01105 Linus introduced gitlinks to update-index. He explains
that he thinks it is not the right thing to replace a gitlink with
something else.

That commit is from the very first beginnings of submodule support.
Since then we have gotten a lot closer to being able to remove a
submodule without losing its history. This check prevents such a use
case, so I think this assumption has changed.

Additionally in the git add codepath we do not have such a check, so for
consistency reasons I think removing this check is the correct thing to
do.

Signed-off-by: Heiko Voigt <hvoigt@hvoigt.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: bf47e238b4c0b1926aa3e357ec9ae795c084ba23 [formerly 413c5625d52048eaaa801b21ee7f9b0a719ab883] [formerly 413c5625d52048eaaa801b21ee7f9b0a719ab883 [formerly 242f55f61252e610092e93f4b2daebba21cb8acf]]
Former-commit-id: aa6ba42b10c02f8a04ecf1c35da4019a777dbd35
Former-commit-id: d568dbdbef08c2d729f45222c611a29b02647eb0