completion: warn people about duplicated function
The __gitdir function is duplicated between completion and prompt
scripts, and these definitions should not diverge; otherwise one of
them can be subtly broken depending on the order the user's shell
dot-sources them.

Leave a note to people who may want to touch one copy to make sure
they update the other one in sync.  Hopefully this line would also
appear in the context of the patch to allow reviewers to notice a
patch that attempts to update only one of them.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5602cd74e4e7f306324844200c6d657aaf806800 [formerly 5602cd74e4e7f306324844200c6d657aaf806800 [formerly ac3eb1c38420b356b775ed63f7291eb4054d11d9]]
Former-commit-id: 84b875c5f54cb82f6b8f5138f3dce6b8579e93d9
Former-commit-id: 5396f9522a46e7b25584e70db09ba5012eb5c346