t: Replace 'perl' by $PERL_PATH
GIT-BUILD-OPTIONS defines PERL_PATH to be used in the test suite. Only a
few tests already actually use this variable when perl is needed. The
other test just call 'perl' and it might happen that the wrong perl
interpreter is used.

This becomes problematic on Windows, when the perl interpreter that is
compiled and installed on the Windows system is used, because this perl
interpreter might introduce some unexpected LF->CRLF conversions.

This patch makes sure that $PERL_PATH is used everywhere in the test suite
and that the correct perl interpreter is used.

Signed-off-by: Vincent van Ravesteijn <vfr@lyx.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 47c8c1ceea709431fe7def7ed28f1406f078312d [formerly 77e5f13e0915d760940d7dbe31a566ebf6736d27] [formerly 77e5f13e0915d760940d7dbe31a566ebf6736d27 [formerly a3428205e6d74542d6441baaa29d1cb1d1064d95]]
Former-commit-id: fa17eba1aacb5f40fd00b3516793961eb38d13fc
Former-commit-id: dd7bff8872105665931d68346c94650c157203a1