t/README: add a bit more Don'ts
Add a few more advices that we often have to give to new test
writers.

Also update an example where a double quote pair is used to enclose
a test body to use a single quote pair, which is more readable and
more importantly gives saner semantics for variable substitution.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b6a992d25eaf06c6593c4df853e57911e5c28153 [formerly b6a992d25eaf06c6593c4df853e57911e5c28153 [formerly ad78585eeebd8d5ac596ad4f06f549734ce9d870]]
Former-commit-id: 845318d465e48af9e6eebdbd3e931c41f3de6b15
Former-commit-id: 2b8d1681628bf9ef30c1c3a2e2e8c61835856d62