gitweb: Skip 'modification times' tests when no date parser available
The If-Modified-Since support in Gitweb is conditional on the
availability of a date parser from either the HTTP::Date or
Time::ParseDate modules. If a suitable parser is not available,
then the corresponding 'modification times' tests should be skipped.

Introduce the DATE_PARSER test prerequisite and use it to skip
all of the dependent tests.

Signed-off-by: Ramsay Jones <ramsay@ramsay1.demon.co.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e9cc29257579fa198b01baa857d416577801e38f [formerly e9cc29257579fa198b01baa857d416577801e38f [formerly 2a8a4490c20e3f39dc15721b49715867511163ba]]
Former-commit-id: c7fd455cc6f2b4d312ad0c070f8b47da60549f72
Former-commit-id: 38d3fe93d5745faf1ccdc9c0671a57c5b17aada7