git: Wrong parsing of ssh urls with IPv6 literals ignores port
If we encounter an address part shaped like "[HOST]:PORT", we skip the opening
bracket and replace the closing one with a NUL.  The variable host then points
to HOST and we've cut off the PORT part.  Thus, when we go looking for it using
host a bit later, we can't find it.  Start at end instead, which either points
to the colon, if present, or is equal to host.

Signed-off-by: Rene Scharfe <rene.scharfe@lsrfire.ath.cx>
Reviewed-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c31a32d4f8e1f52f9555412ea366fe8e28ebc138 [formerly 1d378862f7baaa69ebf98d51437ce444d2955305] [formerly 1d378862f7baaa69ebf98d51437ce444d2955305 [formerly 7acf438215d1b0e8e47a707f3585de8486a2e5fe]]
Former-commit-id: c7cfa1aa7e55d9e3490d14cdd9af489cfb63432e
Former-commit-id: 43dda323bd0ab3d7f72cc1191fe277dfc99dc2f4