apply: factor out checkout_target() helper function
When a patch wants to touch a path, if the path exists in the index
but is missing in the working tree, "git apply --index" checks out
the file to the working tree from the index automatically and then
applies the patch.

Split this logic out to a separate helper function.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d67d7c57e3e7ced8cb8c69973b0cc0a934db085e [formerly 2288797975d994cc73337bf3e1d55105d09bfbf7] [formerly 2288797975d994cc73337bf3e1d55105d09bfbf7 [formerly f4c66eeddd37d3e5f4043df89c2679fbadeb57bd]]
Former-commit-id: eae95fbd52d41bf06e7fec94dfc2437b26305772
Former-commit-id: 8651d52f1f764873e354890dcd33a1ecb89ffb4a