t7400: avoid path mangling issues
A recently introduced test uses an absolute path. But when run on Windows
using the MSYS bash, such a path is mangled into a Windows style path when
it is passed to 'git config'. The subsequent 'test' then compares the
mangled path to the unmangled version and reports a failure.

A path beginning with two slashes denotes a network directory
(//server/share path) and is not mangled. Use that trick to side-step the
issue. Just in case that 'git submodule init' regresses in such a way that
it accesses the URL, use a path name that is unlikely to exist on POSIX
systems, and that cannot be a server name on Windows.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: fac74781bee569f10b49447b5c3a0ea00b30f6c5 [formerly d77a14418d05a3c73ce1d03ae1e5e7245a92d3bb] [formerly d77a14418d05a3c73ce1d03ae1e5e7245a92d3bb [formerly c517e73d0f116f8293f1fbc7f92bb60ea8047b0f]]
Former-commit-id: bc41757390d68d47637eda8397727dc0238996e5
Former-commit-id: 06c5e217af4d4070c0faae5682c9fecec694b3c7