perl/Makefile.PL: warn about duplicate module list in perl/Makefile
Adding or removing a module requires modifying both files to support
builds with and without MakeMaker.  Add a comment to remind patch
authors and reviewers at the crucial moment.

Longer term, it would be nicer to maintain a single list, perhaps in a
separate file used by both build systems.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7b20621eea1907d2bd6f00c348bcf6bf9fd42f2c [formerly 7b20621eea1907d2bd6f00c348bcf6bf9fd42f2c [formerly d22e567770866a0bbe746316561be50b521eb8a5]]
Former-commit-id: 101a3adb9f48aa007c1f5c1424d9398e99c24ce6
Former-commit-id: c5400423092fcbc8470ec522612f528dc5283261