Makefile: split prefix flags from GIT-CFLAGS
Most of the build targets do not care about the setting of
$prefix (or its derivative variables), but will be rebuilt
if the prefix changes. For most setups this doesn't matter
(they set prefix once and never change it), but for a setup
which puts each branch or version in its own prefix, this
unnecessarily causes a full rebuild whenever the branc is
changed.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a2ae9875e718ddaaa9e6980789e9db6c1e418eef [formerly 7c075fbb9de2507e027fc5bc439121a29dd921c4] [formerly 7c075fbb9de2507e027fc5bc439121a29dd921c4 [formerly be1dbd0a933749526eb763fd00cd00c0bdf7c7a6]]
Former-commit-id: a389a782e67f43d2eaf58e64376ba7c3e8d6f51f
Former-commit-id: baf5293c459473423252084402e1c297ee26f549