Makefile: move GIT-VERSION-FILE dependencies closer to use
There is a list of all of the targets which depend on
GIT-VERSION-FILE, but it can be quite far from the actual
point where the targets actually use $(GIT_VERSION). This
can make it hard to verify that each use of $(GIT_VERSION)
has a matching dependency.

This patch moves the dependency closer to the actual build
instructions, which makes verification easier.  This also
fixes the generation of "configure", which did not properly
mark the dependency.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 981c84ed9ec248bad97c544321f9327badfc85a8 [formerly 66079ffc13e195bd3aaefbf434efad2b9f017082] [formerly 66079ffc13e195bd3aaefbf434efad2b9f017082 [formerly 520a6cdce3531867110df012d62d4974376eba1f]]
Former-commit-id: aaa000e2b671a7360149667813132a9ee82a74df
Former-commit-id: 40fa4afd10be846631203e20987886c87890c29f