docs: always define git-relative-html-prefix attribute
Commit fe77b41 introduced a new attribute to let the linkgit macro
create cross-directory HTML references from the technical/ and howto/
subdirectories back to the main documentation. We define that attribute
to "../" on the command-line when building inside those subdirectories,
and otherwise leave it unset under the assumption that it would default
to being blank.  Instead, asciidoc omits the link entirely, leading to
broken documentation. Fix this by defining git-relative-html-prefix to
blank in asciidoc.conf (and an instance on the command-line, when
present, will override it).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b99b33981d9e398d8d174c1f02f029827621db12 [formerly 77e78bdc7269e96778521780110f146fd8261089] [formerly 77e78bdc7269e96778521780110f146fd8261089 [formerly 60475183c0fe76eab2a12fdf840caa0a29217886]]
Former-commit-id: 23f3087cba30321663ead5a8bba400b0fd77eca9
Former-commit-id: f51d1dad1489437a631153a866c3ac7325fc5cdf