diff-no-index: exit(1) if 'diff --quiet <repo file> <external file>' finds changes
When running 'git diff --quiet <file1> <file2>', if file1 or file2
is outside the repository, it will exit(0) even if the files differ.
It should exit(1) when they differ.

This happens because 'diff_no_index' looks at the 'found_changes'
member from 'diff_options' to determine if changes were made.  This
is the wrong thing to do, since it is only set if xdiff is actually
run and it finds a change (the diff machinery will optimize out the
xdiff call when it is not necessary) and in that case HAS_CHANGED
flag needs to be taken into account.

Use diff_result_code() that knows all these details for the correct
exit value instead.

Signed-off-by: Tim Henigan <tim.henigan@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 001aa9b937ba4d939683c7ef21d57d16086aaa59 [formerly b7011dcecc3de599ea935f15907461bcab7514af] [formerly b7011dcecc3de599ea935f15907461bcab7514af [formerly 304970dd5d9e542e70eca98932c1e9f26770678e]]
Former-commit-id: e3737bc6482b22117d8ba211340e7c294846d155
Former-commit-id: dfb1389c69db0e374b9f3a8f5e656f243ff8faed