git credential fill: output the whole 'struct credential'
Instead of outputing only the username and password, print all the
attributes, even those that already appeared in the input.

This is closer to what the C API does, and allows one to take the exact
output of "git credential fill" as input to "git credential approve" or
"git credential reject".

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3fdc6f7004d30d787ccf184a6c69709ab6d91864 [formerly 3fdc6f7004d30d787ccf184a6c69709ab6d91864 [formerly 2d6dc182b8fa171a6b283ce6e8e75a35e13ea67a]]
Former-commit-id: 22baac8b73d8777f8b90650852d1e929624e5580
Former-commit-id: 053760d00acf6c2be2b5f4a22e5e08f373553fb1