git-remote-mediawiki: add credential support
The previous version implemented the possibility to log in a wiki, but
the username and password had to be provided as configuration variables.
We add the possibility to use the Git credential system to prompt
the password.

The support if implemented with generic functions that mimic the C API,
designed to be usable from other contexts in the future (i.e. they may
migrate to Git.pm if someone is interested).

While we're there, do a bit of refactoring in mw_connect_maybe.

Based on patch by: Javier Roucher Iglesias <Javier.Roucher-Iglesias@ensimag.imag.fr>

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4f0e4c92e4e1d1932768bfd1f7e505df03cb5ed5 [formerly 4f0e4c92e4e1d1932768bfd1f7e505df03cb5ed5 [formerly a7271ad1dccaf1a1875992520204debb37ed78bb]]
Former-commit-id: ba4dbdf6512a78354e1702aa2716ee082543eb21
Former-commit-id: fbfb1fc6e3d9f8307d34084409763880abed9b7b