test-lib: reorder and include GIT-BUILD-OPTIONS a lot earlier
This dot-sources GIT-BUILD-OPTIONS a lot earlier in test-lib.sh so
that its use of "perl" can use "$PERL_PATH" to choose the version of
Perl the user told us is suitable for our use.

This is iffy; I didn't check it very carefully, and I would not be
surprised if there are subtle breakages.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1920e1dc7ceeea8bd1e2f0740fd8661939eb61b1 [formerly 1920e1dc7ceeea8bd1e2f0740fd8661939eb61b1 [formerly 3c8f12c96ca09af2a0a9df9fbb9ffd25985ff597]]
Former-commit-id: 9f00ce3b0c71af26307f8c934a4a0966769fa2fb
Former-commit-id: f049be87fe6a98b3d9d1b71412a56c02bb9fbb27