rebase --root: print usage on too many args
Just like

  git rebase --onto newbase upstream branch error

displays the usage message, so should clearly

  git rebase --onto newbase --root branch error

, but it doesn't. Instead, it ignores both "branch" and "error" and
rebases the current HEAD. This is because we try to match the number
of remainging arguments "$#", which fails to match "1" argument and
matches the "*" that really should have been a "0".

Make sure we display usage information when too many arguments are
given. Also fail-fast in case of similar bugs in the future by
matching on exactly 0 arguments and failing on unknown numbers.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: cefb3344cf2675c5c3be3af060a8e92829fc6f25 [formerly fd88925dba317bfbb1522707506ee8bc26af1145] [formerly fd88925dba317bfbb1522707506ee8bc26af1145 [formerly f2b6a19907ccb366790c7ec7b640f9111ac7ad32]]
Former-commit-id: 630eb9bf7f10c12a577457d9de58db1f177761dd
Former-commit-id: f579283256e5b2d7e2c393b8b492498042539ccc