index-pack: Disable threading on cygwin
The Cygwin implementation of pread() is not thread-safe since, just
like the emulation provided by compat/pread.c, it uses a sequence of
seek-read-seek calls. In order to avoid failues due to thread-safety
issues, commit b038a61 disables threading when NO_PREAD is defined.
(ie when using the emulation code in compat/pread.c).

We introduce a new build variable, NO_THREAD_SAFE_PREAD, which allows
use to disable the threaded index-pack code on cygwin, in addition to
the above NO_PREAD case.

Signed-off-by: Ramsay Jones <ramsay@ramsay1.demon.co.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 373efaa3c89f7fc89dc106de7e086a99dd45702a [formerly 373efaa3c89f7fc89dc106de7e086a99dd45702a [formerly c0f86547c522906da7538c0d2b7b0de608499c23]]
Former-commit-id: cc2362c56eb3a66c35a494f2cac72b51f6f33c16
Former-commit-id: b637796e28533400b1d29851365439fd7a466ad2