rebase -i: support --root without --onto
Allow --root to be specified to rebase -i without --onto, making it
possible to edit and re-order all commits right back to the root(s).

If there is a conflict to be resolved when applying the first change,
the user will expect a sane index and working tree to get sensible
behaviour from git-diff and friends, so create a sentinel commit with an
empty tree to rebase onto. Automatically squash the sentinel with any
commits rebased directly onto it, so they end up as root commits in
their own right and retain their authorship and commit message.

Implicitly use rebase -i for non-interactive rebase of --root without
an --onto argument now that rebase -i can correctly do this.

Signed-off-by: Chris Webb <chris@arachsys.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 93694146fa9be03b689862891f5a451ae6ab247a [formerly fb3c681cb8e8f347a82af8752c9545c2f24ff3d3] [formerly fb3c681cb8e8f347a82af8752c9545c2f24ff3d3 [formerly df5df20c1308f936ea542c86df1e9c6974168472]]
Former-commit-id: fadc4e8a49c826cee2828ac45f8557e8496a62e7
Former-commit-id: 6fce2b9ad5af0bd000ff644c248cf0abbbbb20c7