Add tests for rebase -i --root without --onto
Test for likely breakages in t3404, including successful reordering of
non-conflicting changes with a new root, correct preservation of commit
message and author in a root commit when it is squashed with the
sentinel, and presence of the sentinel following a conflicting
cherry-pick of a new root.

Remove test_must_fail for git rebase --root without --onto from t3412 as
this case will now be successfully handled by an implicit git rebase -i.

Signed-off-by: Chris Webb <chris@arachsys.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 817c354fc684d39c12e0711c31aa18532f5b4840 [formerly 817c354fc684d39c12e0711c31aa18532f5b4840 [formerly b64b7feb730b7a23c7a473a484ba6e68d410c990]]
Former-commit-id: 622795f36d197e72b655a8aa36e3be6397e546c6
Former-commit-id: 36003f16f73507403d0075d82cd1c6125c3c0336