git-remote-mediawiki: send "File:" attachments to a remote wiki
The current version of the git-remote-mediawiki supports only import and
export of plain wiki pages. This patch adds the functionality to export
file attachments (i.e. the content of the File: MediaWiki namespace),
which are also exposed by MediaWiki API.

This requires a recent version of MediaWiki::API (Version 0.37 works.
Version 0.34 doesn't).

Signed-off-by: Pavel Volek <Pavel.Volek@ensimag.imag.fr>
Signed-off-by: NGUYEN Kim Thuat <Kim-Thuat.Nguyen@ensimag.imag.fr>
Signed-off-by: ROUCHER IGLESIAS Javier <roucherj@ensimag.imag.fr>
Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d34b0c1623417bf3c833c285e1d589cf68012c7c [formerly d34b0c1623417bf3c833c285e1d589cf68012c7c [formerly b3d985959869d3f2fb1901639326eb64c2e3f726]]
Former-commit-id: 7655c00bd7dad8752348df796aa2d11afb2b6d3a
Former-commit-id: c660e4e987f1ac0d63e315613817ae15a095f328