git p4 test: use real_path to resolve p4 client symlinks
The p4 program is finicky about making sure the recorded client Root
matches the current working directory.  The way it discovers the latter
seems to be to inspect shell variable $PWD.  This could involve symlinks,
that while leading to the same place as the client Root, look different,
and cause p4 to fail.

Resolve all client paths using "test-path-utils real_path $path".  This
removes ".." and resolves all symlinks.

Discovered while running with --root=/dev/shm, which is a link to
/run/shm.

Signed-off-by: Pete Wyckoff <pw@padd.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1dd160f5cedfaef6854e3a9e6b357306ff3a207d [formerly 1dd160f5cedfaef6854e3a9e6b357306ff3a207d [formerly 23bd0c99f7b789077c40ae5384dd0054ef5c00ac]]
Former-commit-id: f298e81b606502062bd6517777abc5799cb01c3c
Former-commit-id: 42f5851827a8f62173991bdcdfb9d63085d0ebde