git p4 test: split up big t9800 test
The original t9800 test code has a mix of assorted topics, some
of which are big enough to deserve their own homes.
Interdependencies between the topics make it confusing when
trying to study one in isolation.  And it takes so long to run
that debugging an individual test is difficult.

Split out three big chunks of tests into their own files:

    t9812-git-p4-wildcards.sh gets the 8 p4 wildcard tests

    t9813-git-p4-preserve-users.sh gets the 4 --preserve-user tests

    t9814-git-p4-rename.sh gets the 2 copy and rename tests

Test 9800 execution time drops from 29 sec to 9 sec.  The
sequential time to run all tests is a slower due to the three
extra p4d startup/shutdown sequences, but the overall parallel
execution time is about the same, at 52 sec.

Signed-off-by: Pete Wyckoff <pw@padd.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 540a3586a4999198da0d5c28497879754ff0e4e0 [formerly 540a3586a4999198da0d5c28497879754ff0e4e0 [formerly 9b6513ac6fde1552a9a3d606be704eb5ded43ac0]]
Former-commit-id: cc098eae5d494181f6ba3613246a447bff0fc5ce
Former-commit-id: 279ea27b8f351794762994a523c99002ae7ebca6