git p4 test: fix badp4dir test
The construct used to get the return code was flawed, in that
errors in the &&-chain before the semicolon were not caught.  Use
the standard test_expect_code instead.

Set PATH in a subshell instead of relying on the bashism of
setting it just for a single command.

And fix the grep line so it doesn't worry about grep segfaults,
and doesn't fail for i18n issues.

Reported-by: Johannes Sixt <j.sixt@viscovery.net>
Signed-off-by: Pete Wyckoff <pw@padd.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: aca2f89814a31708fdc36702499899c738ca403b [formerly aca2f89814a31708fdc36702499899c738ca403b [formerly a0327c0edc7ada60ae2af5ec8daa5dcfcacd095d]]
Former-commit-id: 8f27f7b749b422e81b0a5eb3c2098c6cf808e523
Former-commit-id: ec52f55859322e31ec882e4167de132847ce3169