git p4 test: wait longer for p4d to start and test its pid
Running tests at high parallelism on a slow machine, 5 sec is
not enough to wait for p4d to start.  Change it to 5 minutes,
adding an environment variable P4D_START_PATIENCE to shrink
that if needed in automated test environments.

Also check if the pid of the p4d that we started is still
around.  If not, quit waiting for it immediately.

Remove all the confusing && chaining and simplify the code.

Thanks-to: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Pete Wyckoff <pw@padd.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ae6cb6097f1fb2ec7b25c7b012d7a46aaed77651 [formerly 4640ea005c5de6585406c601abff725a64c010ee] [formerly 4640ea005c5de6585406c601abff725a64c010ee [formerly f89f35a9d41535b3e8ebe500489634e623ae8d83]]
Former-commit-id: 3525c8e138b1a976bd7bd49eeec77bb8ef6147f1
Former-commit-id: 14ff5e1b5fcc0a8609060dc7dee5cc3584a5f016