commit/commit-tree: correct latin1 to utf-8
When a line in the message is not a valid utf-8, "git mailinfo"
attempts to convert it to utf-8 assuming the input is latin1 (and
punt if it does not convert cleanly).  Using the same heuristics in
"git commit" and "git commit-tree" lets the editor output be in
latin1 to make the overall system more consistent.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3ca28e044bfd1415ccce579264f394c1764a5bf3 [formerly 3ca28e044bfd1415ccce579264f394c1764a5bf3 [formerly 08a94a145c3231c0fa36469682591a3c45222271]]
Former-commit-id: 0fc41739d3b3dcf8162920a450354a35df78faf0
Former-commit-id: cd16ba17f5c4476cdb9f907d5a0e32d5d3846f01