gitweb: Cleanup git_print_log()
When we see a signed-off-by line (and its friends), we set $signoff
to true, but then we process the next line after we are done without
giving control to the rest of the loop.  And when the line we saw is
not a signed-off-by line, we reset $signoff to false before running
the remainder of the loop.

Hence, the check for $signoff that attempts to remove an extra empty
line between two signed-off-by line was not doing anything useful.

Rename $empty to a more explicit name $skip_blank_line to tell us to
skip a blank line when we see one, set it after we see and emit a
blank line (to avoid showing more than one empty lines in a raw) or
after we handle a signed-off-by line (to avoid empty lines after
such a line), to fix this bug, and get rid of the $signoff variable
that is not useful.

Signed-off-by: Namhyung Kim <namhyung@kernel.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c9bfa4ad11341b6568b6090548ce91703a2a2158 [formerly bee918c332a00870c8ed2b65c13b1db32ba516f8] [formerly bee918c332a00870c8ed2b65c13b1db32ba516f8 [formerly 5a45c0cafe6d73f6026bd0ea525da2f9765b112d]]
Former-commit-id: 727b918c22eb35809e9dc1891fbffc258ef5a53e
Former-commit-id: 50d64462c35589a0dd39dbb563403f757493640c