t3404: make test 57 work with dash and others
The construct

	VAR=value test_must_fail command args

works only for some shells (such as bash) but not others (such as dash)
because VAR=value does not end up in the environment for command when it
is called by the shell function test_must_fail. That is why we explicitly
set and export variable in a subshell, i.e.

	(
		VAR=value &&
		export VAR &&
		test_must_fail command args
	)

in most places already, bar the newly introduced 57 from b64b7fe
(Add tests for rebase -i --root without --onto, 2012-06-26).

Make test 57 use that construct also.

Signed-off-by: Michael J Gruber <git@drmicha.warpmail.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a44cdcabe6af45b8c7ca96e54638ee8209dfeae0 [formerly e83599737be38f61b3da4d528c94b5fd3ad585fc] [formerly e83599737be38f61b3da4d528c94b5fd3ad585fc [formerly 994fd91d1fd5e9feffc88ab7db19e9c9461b4543]]
Former-commit-id: f882d99fb564091213e43e3b7e9b4a329338d5ec
Former-commit-id: 14d7b9988104a432cea44bbfb9de843ff536543e