git p4: notice Jobs lines in git commit messages
P4 has a feature called "jobs" that allows linking changes
to a bug tracking system or other tasks.  When submitting
code, a job name can be specified to mark that this change
is associated with a particular job.

Teach git-p4 to find an optional "Jobs:" line in git commit
messages and use them to make a Jobs section in the p4
change specifitation.

Signed-off-by: Pete Wyckoff <pw@padd.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3968e2f5c88d7a9ead4a21ff718faf37e14c2a9a [formerly d7e850512bd5c375d2151e4a3f24564e34cc2959] [formerly d7e850512bd5c375d2151e4a3f24564e34cc2959 [formerly f19cb0a0e80e094970cfa854dc3da60c927830eb]]
Former-commit-id: dc4708183ea20e0508883b452d852f68e9feb939
Former-commit-id: 8003920680b36e2ea5813aa172a1ac5d71b3b2fe