git-remote-mediawiki: test environment of git-remote-mediawiki
In order to test git-remote-mediawiki, a set of functions is needed to
manage a MediaWiki: edit a page, remove a page, fetch a page, fetch all
pages on a given wiki.

A few helper function are also provided to check the content of
directories.

In addition, this patch provides Makefiles to execute tests.
See the README file for more details.

Signed-off-by: Simon CATHEBRAS <Simon.Cathebras@ensimag.imag.fr>
Signed-off-by: Julien KHAYAT <Julien.Khayat@ensimag.imag.fr>
Signed-off-by: Simon Perrat <simon.perrat@ensimag.imag.fr>
Signed-off-by: Charles ROUSSEL <Charles.Roussel@ensimag.imag.fr>
Signed-off-by: Guillaume SASDY <Guillaume.Sasdy@ensimag.imag.fr>
Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d7e8aef55b8f5403e0937d6f0e530906ce03b6cb [formerly 85742e98db40dbcbbe2c30e57015fcb18bb4f2cd] [formerly 85742e98db40dbcbbe2c30e57015fcb18bb4f2cd [formerly 8435b289890739600436b6c902156fc073752ffb]]
Former-commit-id: bf97313663db3bf66c280cd3910ec17e3f5c0dae
Former-commit-id: d0e1ef0b645a76ff14a557bea6ecdd11e0858210