Makefile: document ground rules for target-specific dependencies
When a source file makes use of a makefile variable, there should be a
corresponding dependency on a file that changes when that variable
changes to ensure the build output is not left stale when the variable
changes.

Document this, even though we are not following the rule perfectly
yet.  Based on an explanation from Jeff King.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c0db358758d28e8b5096f6f36650f31a40f80b7d [formerly 7b57583341c93a05834ceb8735548de9f669353c] [formerly 7b57583341c93a05834ceb8735548de9f669353c [formerly 7b63c77eeea7cc4bf3d29793bcb95dcf659e5126]]
Former-commit-id: 408fd6221451d224abc4565ec2e4ffee39eab94a
Former-commit-id: 46586c7b76382eb9d5427e3203727edd624349fd