Makefile: be silent when only GIT_USER_AGENT changes
To avoid noise during builds, unlike the GIT-CFLAGS rule which prints
"* new build flags or prefix" so the operator knows why all files are
being rebuilt when it changes, GIT-USER-AGENT generation is silent.

If this code breaks and a target depending on GIT-USER-AGENT ends up
being rebuilt when it shouldn't be, the full dependency chain can be
retrieved with "make --debug=b".

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7572e8847ce43b5b8bc4f51b9fc78de40bb491fc [formerly 7572e8847ce43b5b8bc4f51b9fc78de40bb491fc [formerly 33ddbcb01208f93c20ac4b904ad3d76777144882]]
Former-commit-id: 077ab04841ee1faa15e1ac4d451b09e53cb0b1a6
Former-commit-id: fa9daea1eb4f65378f9950de3816ba7f3370963e