Strip namelen out of ce_flags into a ce_namelen field
Strip the name length from the ce_flags field and move it
into its own ce_namelen field in struct cache_entry. This
will both give us a tiny bit of a performance enhancement
when working with long pathnames and is a refactoring for
more readability of the code.

It enhances readability, by making it more clear what
is a flag, and where the length is stored and make it clear
which functions use stages in comparisions and which only
use the length.

It also makes CE_NAMEMASK private, so that users don't
mistakenly write the name length in the flags.

Signed-off-by: Thomas Gummerer <t.gummerer@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 21926a7a7b9faad7af3068e5ecb5ac2967d0e45a [formerly 21926a7a7b9faad7af3068e5ecb5ac2967d0e45a [formerly b60e188c51242b72061b5f2f0d4df80397f6125a]]
Former-commit-id: 344f4a98568be7e580f4775c9cda4332df188fe1
Former-commit-id: 6a13f4b1466ea4ec91faf13eef657f1b7450b5f3