revision: avoid work after --max-count is reached
During a revision traversal in which --max-count has been
specified, we decrement a counter for each revision returned
by get_revision. When it hits 0, we typically return NULL
(the exception being if we still have boundary commits to
show).

However, before we check the counter, we call get_revision_1
to get the next commit. This might involve looking at a
large number of commits if we have restricted the traversal
(e.g., we might traverse until we find the next commit whose
diff actually matches a pathspec).

There's no need to make this get_revision_1 call when our
counter runs out. If we are not in --boundary mode, we will
just throw away the result and immediately return NULL. If
we are in --boundary mode, then we will still throw away the
result, and then start showing the boundary commits.
However, as git_revision_1 does not impact the boundary
list, it should not have an impact.

In most cases, avoiding this work will not be especially
noticeable. However, in some cases, it can make a big
difference:

  [before]
  $ time git rev-list -1 origin Documentation/RelNotes/1.7.11.2.txt
  cef7ea5214507f3e17429dfb925e08251f668ba2 [formerly ff24c327c9a867dd129478566f54925e9877438f] [formerly dcbdecbd4d724a25d584578796203ec6dd4cd7bf] [formerly ec51b91cdb0207c1267c42f2610cec13eb7ec582 [formerly ec51b91cdb0207c1267c42f2610cec13eb7ec582 [formerly 8d141a1d562abb31f27f599dbf6e10a6c06ed73e]]]

  real    0m0.301s
  user    0m0.280s
  sys     0m0.016s

  [after]
  $ time git rev-list -1 origin Documentation/RelNotes/1.7.11.2.txt
  cef7ea5214507f3e17429dfb925e08251f668ba2 [formerly ff24c327c9a867dd129478566f54925e9877438f] [formerly dcbdecbd4d724a25d584578796203ec6dd4cd7bf] [formerly ec51b91cdb0207c1267c42f2610cec13eb7ec582 [formerly ec51b91cdb0207c1267c42f2610cec13eb7ec582 [formerly 8d141a1d562abb31f27f599dbf6e10a6c06ed73e]]]

  real    0m0.010s
  user    0m0.008s
  sys     0m0.000s

Note that the output is produced almost instantaneously in
the first case, and then git uselessly spends a long time
looking for the next commit to touch that file (but there
isn't one, and we traverse all the way down to the roots).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d4e05a4a3b46d261255e0fec54eb79edf185497d [formerly d4e05a4a3b46d261255e0fec54eb79edf185497d [formerly b72a1904aefb9e27014a11790f3af4dc90b38e8d]]
Former-commit-id: d24156fddc3cd671351dd8adad6c5511ae8bf606
Former-commit-id: 694aff89b93b7e0acb093de4f830e0f0b6a16bee