status: color in-progress message like other header messages
The "status" command recently learned to describe the
in-progress operation in its long output format (e.g.,
rebasing, am, etc). This message gets its own slot in the
color table, even though it is not configurable. As a
result, if the user has set color.status.header to a
non-default value, this message will not match (and cannot
be made to match, as there is no config option).

It is probably more sane to just color it like the rest of
the text (i.e., just use color.status.header). This would
not allow users to customize the color of this message
independently, but they cannot do that with the current code
anyway, and if somebody wants to build customizable
colorization later, this patch does not make it much harder
to do so.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d17fc5d93f119cfc2c5461b9937db66b3efc8f9e [formerly d17fc5d93f119cfc2c5461b9937db66b3efc8f9e [formerly 3cd9d1b2f4471211a9c620d6ea8fe736745c3f5b]]
Former-commit-id: b38fd265718e50632c306bb9d2fc42c4a2ace578
Former-commit-id: 280a4071d923e6e5fdc326ff58bb452fff52c556