git-remote-mediawiki: make mediafiles export optional
It is possible to use git-remote-mediawiki on a tree with both .mw files
and other files. Before git-remote-mediawiki learnt how to export
mediafiles, such mixed trees allowed the user to maintain both the wiki
and other files for the same project in the same repository. With the
newly added support for exporting mediafiles, pushing such mixed trees
would upload unrelated files as mediafiles, which may not be desired.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 03fbb1342dc9b1553454ed4c28db52eb08d9c393 [formerly f45817c7da7fc09f90861e039efe873e1c791491] [formerly f45817c7da7fc09f90861e039efe873e1c791491 [formerly 2045e293eb69204ac06910c7570718efb09b7059]]
Former-commit-id: a0431607ec9a135ca2ed25b579669eb761cdf4c5
Former-commit-id: 4009fcf32ac79c93da8ff0396e2e5644b5136d69