t4012: Actually quote the sed script
The nested quoting is not needed in this cases, thus the previous
version did work just fine. Nevertheless the usage is misleading,
so just achieve nested quoting by using double quotes instead. Lower
the probability of breakage in the future and make the code easier
to read.

NOTE: Just dropping the single quotes around the sed arguments would
      have also been possible.

Signed-off-by: Alexander Strasser <eclipse7@gmx.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 997a14484dd2f5e911baf57bd9609e7209f6e197 [formerly 997a14484dd2f5e911baf57bd9609e7209f6e197 [formerly 4a2b716e220f0f5369c386d7257b87399d424a3f]]
Former-commit-id: 88bbec16a65b20acefc199b17f7715f6d9e9a281
Former-commit-id: 1a8effd4442e2e243041a71ed4e89965b07bc006