t4012: use 'printf' instead of 'dd' to generate a binary file
For some reason, 'echo X | dd bs=1k seek=1' creates a file with 2050 bytes
on Windows instead of the expected 1026 bytes, so that a test fails. Since
the actual contents of the file are irrelevant as long as there is at
least one zero byte so that the diff machinery recognizes it as binary,
use printf to generate it.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Alexander Strasser <eclipse7@gmx.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 48ebc2577fae8b691f4435d9cd1e4dc527224a3d [formerly 9a341011872c8e9dd753e9899733dd3f56b852f8] [formerly 9a341011872c8e9dd753e9899733dd3f56b852f8 [formerly 66fd93ee060a1e7f831316f6c81e57f43c5578ba]]
Former-commit-id: 061bcf9e5fb2807f4e67dc1bc3c8db7dc908bc8a
Former-commit-id: 8062e4f364993256f392633376e88aa612888f2b