git-remote-mediawiki: allow page names with a ':'
Traditionnally, pages named Foo:Bar are page 'Bar' in namespace 'Foo'.
However, it is also possible to call a page Foo:Bar if 'Foo' is not a
namespace. In this case, the actual name of the page is 'Foo:Bar', in the
main namespace. Since we can't tell with only the filename, query the
wiki for a namespace 'Foo' in these cases, but deal with the case where
no such namespace is found.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 076423c43ac5c71ca1635261f654faec88ff6452 [formerly 076423c43ac5c71ca1635261f654faec88ff6452 [formerly 2272400c93332629e8c34dcf3b1207159352e539]]
Former-commit-id: 07e9acce69e3827945b27736753f15c277608239
Former-commit-id: 2c7d79abfed363128b1eff0d44e3edd366384499