docs/credential: minor clarity fixups
The text in git-credential(1) was copied from
technical/api-credentials, so it still talks about the
input/output format as coming from git to the helper. Since
the surrounding text already indicates that this format is
used for reading and writing with git credential, we can
just remove the extraneous confusing bits.

Signed-off-by: Jeff King <peff@peff.net>
Acked-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7e528f3ad23bd0d925d21b02b40ef8e7f9d9789b [formerly 7e528f3ad23bd0d925d21b02b40ef8e7f9d9789b [formerly 3e5f29e8963a9a6f1774add194ec19583c19fbb6]]
Former-commit-id: f596580d7b2a7827519cd6eada9a3abe36956d5d
Former-commit-id: 3f6d3385e7873187b667a2610a9028810fe2c703