mw-to-git: check blank credential attributes via length
When writing a credential to git-credential, we omit fields
that do not have a true value. This will skip empty or
undefined fields (which we want), but will also accidentally
skip usernames or passwords which happen to have a non-true
value (e.g., "0"). Be more careful by checking for non-zero
length.

Signed-off-by: Jeff King <peff@peff.net>
Acked-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ecf6670804ce96e5c13739eed06161d3cb688168 [formerly ecf6670804ce96e5c13739eed06161d3cb688168 [formerly 6319a2a148adaa78f11414e914f8c9f3f1c52b5e]]
Former-commit-id: 41e3d1db869c3b0ace4c9b498908adf9eaee3833
Former-commit-id: 7b4ffcf223fc210405ee6b88c43c46cc16326f82