build: "make clean" should not remove configure-generated files
Those filed hold variables, settings and information set by the
configuration process run by './configure'; in Autotools-based
build system that kind of stuff should only be removed by
"make distclean".  Having it removed by "make clean" is not only
inconsistent, but causes real confusion for that part of the Git
audience that is used to the Autotools semantics; for example,
an autotools old-timer that has run:

    ./configure --prefix /opt/git

in the past, without running "make distclean" afterwards, would
expect a "make install" issued after a "make clean" to rebuild and
install git in '/opt/git'; but with the current behaviour, the
"make clean" invocation removes (among the other things) the file
'config.mak.autogen', so that the "make install" falls back to the
default prefix of '$HOME', thus installing git in the user's home
directory -- definitely unexpected.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 06d4c9cd3b19548870637020f2f206052cebb205 [formerly 06d4c9cd3b19548870637020f2f206052cebb205 [formerly dc7ace5f776f30fe4e10d8a3fa8146ad21893544]]
Former-commit-id: 368fa74ee98607f16394cf3733d7d0ea5cf3dbb0
Former-commit-id: 28faf9c316bcdc8e74f0fa82001dd99a16c72384