build: reconfigure automatically if configure.ac changes
This provides a reduced but still useful sibling of the Automake's
"automatic Makefile rebuild" feature.  It's important to note that
we take care to enable the new rules only if the tree that has already
be configured with './configure', so that users relying on manual
configuration won't be negatively impacted.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 995043274f82527e1035c41b3065c06c089e6c1a [formerly 995043274f82527e1035c41b3065c06c089e6c1a [formerly 8242ff470f2a34b5c0f193c4010cbf3d9b57f9bd]]
Former-commit-id: 776acb98b9634657aba86850515926a14794d95d
Former-commit-id: f419b9946645e920cac70d7071dbb05a29e6c557