difftool: only copy back files modified during directory diff
When 'difftool --dir-diff' is used to compare working tree files,
it always copies files from the tmp dir back to the working tree
when the diff tool is closed, even if the files were not modified
by the diff tool.

This causes the file timestamp to change. Files should only be
copied from the tmp dir back to the working copy if they were
actually modified.

Signed-off-by: Tim Henigan <tim.henigan@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 203a6b36f6ce19c0574bc9db85f6abdf6dd17f94 [formerly 3699e89905e3ea8f4be453140c27ec159eed8540] [formerly 3699e89905e3ea8f4be453140c27ec159eed8540 [formerly 05df532655e14ba351b4ee8a0acd8f190ed0b745]]
Former-commit-id: 91ebf0e73e00a6cf86dc373a5b3ff5ad5338d00c
Former-commit-id: dd9493db5adfaa111019e873141897bf19741c25