difftool: Use symlinks when diffing against the worktree
Teach difftool's --dir-diff mode to use symlinks to represent
files from the working copy, and make it the default behavior
for the non-Windows platforms.

Using symlinks is simpler and safer since we do not need to
worry about copying files back into the worktree.
The old behavior is still available as --no-symlinks.

Signed-off-by: David Aguilar <davvid@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: cc607e793be28e939db717bf80c9dd23fa69ecee [formerly cc607e793be28e939db717bf80c9dd23fa69ecee [formerly 1f2293457579a5bf22bb9c8324ded22b10705cc2]]
Former-commit-id: 0c84bba40cf0f19f5aa34190c0ca1eb43ec5ddc6
Former-commit-id: aa7a9db9856c8d937168a77abc807f4000570b4a