mergetool: support --tool-help option like difftool does
This way we do not have to risk the list of tools going out of sync
between the implementation and the documentation.

In the same spirit as bf73fc2 (difftool: print list of valid tools
with '--tool-help', 2012-03-29), trim the list of merge backends in
the documentation.  We do not want to have a complete list of valid
tools; we only want a list to help people guess what kind of things
the tools do to be specified there, and refer them to --tool-help
for a complete list.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2da4e312263f98992d9ada2aa31eea7e65758d1b [formerly 755a67906015ecc4e21e58c418df3847c6d08a09] [formerly 755a67906015ecc4e21e58c418df3847c6d08a09 [formerly 109859e27445a3ad12eb8f8cba43a5aa8f701733]]
Former-commit-id: 2237a39262e61867dad07e19f3485c9574e37026
Former-commit-id: 52f20ea17b2de6fbc700c00e857e258e546858e3