attr: make sure we have an xdg path before using it
If we don't have a core.attributesfile configured, we fall
back to checking XDG config, which is usually
$HOME/.config/git/attributes.

However, if $HOME is unset, then home_config_paths will return
NULL, and we end up calling fopen(NULL).

Depending on your system, this may or may not cause the
accompanying test to fail (e.g., on Linux and glibc, the
address will go straight to open, which will return EFAULT).
However, valgrind will reliably notice the error.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 71485397a4076008c31a0cdc8681d1386ca269ee [formerly f1159f86e23553fef26dc2e9dff2c73f6c7c258d] [formerly f1159f86e23553fef26dc2e9dff2c73f6c7c258d [formerly f0c1c15c41bdcdaf71c69355ac83789466820879]]
Former-commit-id: 86efbd6e5b06142ae3d80d45122ec348c8c64e43
Former-commit-id: ca4370dbf31ed66fe0101573bd3b1f33784c7469