rebase -i: handle fixup of root commit correctly
There is a bug with git rebase -i --root when a fixup or squash line is
applied to the new root. We attempt to amend the commit onto which they
apply with git reset --soft HEAD^ followed by a normal commit. Unlike a
real commit --amend, this sequence will fail against a root commit as it
has no parent.

Fix rebase -i to use commit --amend for fixup and squash instead, and
add a test for the case of a fixup of the root commit.

Signed-off-by: Chris Webb <chris@arachsys.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 754006354156359d54e096562905316b8407db41 [formerly 754006354156359d54e096562905316b8407db41 [formerly 2147f844ed13fa5052161d38f8cf7dca6f83c06e]]
Former-commit-id: cb6ebade43fe71b98fc83b760886a5481a4b6565
Former-commit-id: 1b06cef3db86b8cd3ba6c5b0f374fc9bfbde5f24