t/lib-httpd: handle running under --valgrind
Running the http tests with valgrind does not work for two
reasons:

  1. Apache complains about following the symbolic link from
     git-http-backend to valgrind.sh.

  2. Apache does not pass through the GIT_VALGRIND variable
     to the backend CGI.

This patch fixes both problems. Unfortunately, there is a
slight hack we need to handle passing environment variables
through Apache. If we just tell it:

  PassEnv GIT_VALGRIND

then Apache will complain when GIT_VALGRIND is not set. If
we try:

  SetEnv GIT_VALGRIND ${GIT_VALGRIND}

then when GIT_VALGRIND is not set, it will pass through the
literal "${GIT_VALGRIND}". Instead, we now unconditionally
pass through GIT_VALGRIND from lib-httpd.sh into apache,
even if it is empty.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5c65c7265d9c9af36e90f252a5b1057a2ef25db7 [formerly 1f7e066e13b8bad011618f9a0337e807a37bdc7c] [formerly 1f7e066e13b8bad011618f9a0337e807a37bdc7c [formerly f628825481b706baa3d9bf871a081cb1285c2778]]
Former-commit-id: eb93c8c1fa0d38ee01378dc1c9084e9409992075
Former-commit-id: 979dfc588f43611a0020417bf405e03dfd53975f