difftool: Handle finding mergetools/ in a path with spaces
Use the original File::Find implementation from bf73fc2 (difftool:
print list of valid tools with '--tool-help', 2012-03-29) so that we
properly handle mergetools/ being located in a path containing
spaces.

One small difference is that we avoid using a global variable by
passing a reference to the list of tools.

Signed-off-by: David Aguilar <davvid@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4b0297b7cba94a14f5fb4e1de3737467718883a2 [formerly 4b6ae933eef3ab3be294b7d5778c46104caebbcb] [formerly 4b6ae933eef3ab3be294b7d5778c46104caebbcb [formerly 7c7584b97096a168fe1236c84e5e12d7bee24476]]
Former-commit-id: 3a402c02631fcad1fd692598432f0844a824b0a2
Former-commit-id: 44c405e70b45a01a3bcc3b09f720f9595e9036e9