i18n: rebase: mark messages for translation
Mark messages in git-rebase.sh for translation.  While doing this
Jonathan noticed that the comma usage and sentence structure of the
resolvemsg was not quite right, so correct that and its cousins in
git-am.sh and t/t0201-gettext-fallbacks.sh at the same time.

Some tests would start to fail with GETTEXT_POISON turned on after
this update.  Use test_i18ncmp and test_i18ngrep where appropriate
to mark strings that should only be checked in the C locale output
to avoid such issues.

Signed-off-by: Jiang Xin <worldhello.net@gmail.com>
Reviewed-by: Stefano Lattarini <stefano.lattarini@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e7812bcd8be314f411347d72dcd5c0391f952ad3 [formerly e7812bcd8be314f411347d72dcd5c0391f952ad3 [formerly c7108bf9ed72403bdf6274bf46263731590bffcf]]
Former-commit-id: 8d84431d7b3bf27bc6b5ca69e5747ce0a56ae17a
Former-commit-id: 2721b169f441b57e8d1eef4cf3561db1399c8e73