difftool: Handle compare() returning -1
Keep the temporary directory around when compare()
cannot read its input files, which is indicated by -1.

Defer tempdir creation to allow an early exit in setup_dir_diff().
Wrap the rest of the entry points in an exit_cleanup() function
to handle removing temporary files and error reporting.

Print the temporary files' location so that the user can
recover them.

Signed-off-by: David Aguilar <davvid@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 07a31716eb01ea42c58eea565fb75a4fc97130db [formerly 07a31716eb01ea42c58eea565fb75a4fc97130db [formerly ceb1497a74e75a0c9bd53716b787d33ad6e2397f]]
Former-commit-id: 2699602473a5d74d6d989cca8bd608e3bf677f53
Former-commit-id: 19c1acf240a09fb4d78ba635846df5fb19506881