t7502: clean up fake_editor tests
Using write_script saves us a few lines of code, and means
we consistently use $SHELL_PATH.

We can also drop the setting of the $pwd variable from
$(pwd). In the first instance, there is no reason to use it
(we can just use $(pwd) directly two lines later, since we
are interpolating the here-document). In the second
instance, it is totally pointless and probably just a
cut-and-paste from the first instance.

Finally, we can use a non-interpolating here document for
the final script, which saves some quoting.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: da52b37c2715023ee110c4a0b065ef2b7315e706 [formerly da52b37c2715023ee110c4a0b065ef2b7315e706 [formerly a9ebc43bd07e219922d1dab17df319f1cae5d600]]
Former-commit-id: da26c32616a70c4cdcb85afd1fa90165b5543ab8
Former-commit-id: 77705c4603bf41d9d1f64c7a6cfd4f988c6cd452