t7502: properly quote GIT_EDITOR
One of the tests tries to ensure that editor is not run due
to an early failure. However, it needs to quote the pathname
of the trash directory used in $GIT_EDITOR, since git will
pass it along to the shell. In other words, the test would
pass whether the code was correct or not, since the unquoted
editor specification would never run.

We never noticed the problem because the code is indeed
correct, so git-commit never even tried to run the editor.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: bbd1cfddb2831a46fbb9d2a19eba92a83b8a75ca [formerly bbd1cfddb2831a46fbb9d2a19eba92a83b8a75ca [formerly 34565f27fef10546a62c734c55c0d280aad63460]]
Former-commit-id: 788b2744f1af49054a9dd0e915a8e89c170c301a
Former-commit-id: df7a47972e7f8a9f44327e9217e3b6cd845dc10a