remove unnecessary parameter from get_patch_ids()
get_patch_ids() takes an already initialized rev_info and a
prefix. The prefix is used when initalizing a second rev_info. Since
the initialized rev_info already has a prefix and the prefix never
changes, we can use the prefix from the initialized rev_info to
initialize the second rev_info.

Signed-off-by: Martin von Zweigbergk <martin.von.zweigbergk@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 981876565065c727ce8a87ba615ee80858dc369f [formerly 981876565065c727ce8a87ba615ee80858dc369f [formerly 811929f1695ed7de15cad14b9335534af672b472]]
Former-commit-id: da3175bbd1083dc052c96a50f2635fb2de9473f1
Former-commit-id: c4350adeecb3ae0f9821b41de13260a65b1f53b3