git-svn: attempt to mimic SVN 1.7 URL canonicalization
Previously, our URL canonicalization didn't do much of anything.
Now it actually escapes and collapses slashes.  This is mostly a cut & paste
of escape_url from git-svn.

This is closer to how SVN 1.7's canonicalization behaves.  Doing it with
1.6 lets us chase down some problems caused by more effective canonicalization
without having to deal with all the other 1.7 issues on top of that.

* Remote URLs have to be canonicalized otherwise Git::SVN->find_existing_remote
  will think they're different.

* The SVN remote is now written to the git config canonicalized.  That
  should be ok.  Adjust a test to account for that.

[ew: commit title]

Signed-off-by: Eric Wong <normalperson@yhbt.net>

Former-commit-id: 45391b5a18f43e55586eee7f58a8229750e22f90 [formerly 45391b5a18f43e55586eee7f58a8229750e22f90 [formerly 93c3fcbe4d4893fac6c9de64219b2eda0b309a13]]
Former-commit-id: d23aadb913168ef66114514d3ef94d4ffc0f6480
Former-commit-id: 51f43fd7dae9e612cbf6c17e1e54fc4a5f6b9237