t1100-*.sh: Fix an intermittent test failure
In particular, the final test ('flags and then non flags') fails
intermittently, depending on how much time elapsed between the
invocations of "git commit-tree" when creating the commits which
later have their commit id's compared. For example, if the commits
for childid-3 and childid-4 are created 1 or more seconds apart,
then the commits, which would otherwise be identical, will have
different commit id's.

In order to make the test reproducible, we remove the variability
by setting the author and committer times to a well defined state.
We accomplish this with a single call to 'test_tick' at the start
of the test.

Signed-off-by: Ramsay Jones <ramsay@ramsay1.demon.co.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2924e5ecdc7c611375bb8f30fc79b8be97f62eeb [formerly 2924e5ecdc7c611375bb8f30fc79b8be97f62eeb [formerly 425b8a22aa14d624c807d5eed92c92f5cd32f376]]
Former-commit-id: 55bbfec689e4e45aeb1618bbf899960c6fea45e3
Former-commit-id: 9374136d9ba051e0dee1cf7fda12b428b241604a