setup: clarify error messages for file/revisions ambiguity
The previous "Use '--' to separate filenames from revisions" may sound
obvious for an old-time Unix user, but does not make it clear how to use
this '--'. In addition to mentionning this '--', give an idea of what the
new command should look like.

Ideally, we could provide cut-and-paste ready commands based on the
command that just failed, but we have no easy access to argv[] in this
place of the code.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 98e90a99e14583707ea1bfb61a1ec1166b0d3137 [formerly 98e90a99e14583707ea1bfb61a1ec1166b0d3137 [formerly 4d4b573977833f887635be386b84681e73834b72]]
Former-commit-id: d9f64943f2bf83be843339b0c60dd01e362f68ac
Former-commit-id: 835ade7e9259c86ee1736d0a5d84438df5c5a9ba