include agent identifier in capability string
Instead of having the client advertise a particular version
number in the git protocol, we have managed extensions and
backwards compatibility by having clients and servers
advertise capabilities that they support. This is far more
robust than having each side consult a table of
known versions, and provides sufficient information for the
protocol interaction to complete.

However, it does not allow servers to keep statistics on
which client versions are being used. This information is
not necessary to complete the network request (the
capabilities provide enough information for that), but it
may be helpful to conduct a general survey of client
versions in use.

We already send the client version in the user-agent header
for http requests; adding it here allows us to gather
similar statistics for non-http requests.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f33d0e4c1fcd01474de9973be87bf215b85a5bb3 [formerly b07f7e30dc11c78c977e00f622bbba9f8005099f] [formerly b07f7e30dc11c78c977e00f622bbba9f8005099f [formerly ff5effdf455bd6e694ff0bb3e0793515a2214adb]]
Former-commit-id: 4156ba84b414b236b153e077afee00da33903814
Former-commit-id: c558304b68999f4d7db8293e21e00363764d60b7