Fix '\ No newline...' annotation in rewrite diffs
When a file that ends with an incomplete line is expressed as a
complete rewrite with the -B option, git diff incorrectly
appends the incomplete line indicator "\ No newline at end of
file" after such a line, rather than writing it on a line of its
own (the output codepath for normal output without -B does not
have this problem).  Add a LF after the incomplete line before
writing the "\ No newline ..." out to fix this.

Add a couple of tests to confirm that the indicator comment is
generated on its own line in both plain diff and rewrite mode.

Signed-off-by: Adam Butcher <dev.lists@jessamine.co.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 01dcfcbb658c0309af5bda46b02cf96a04d26195 [formerly 01dcfcbb658c0309af5bda46b02cf96a04d26195 [formerly 35e2d03c2c24462717d774cd9f4eea46f1474ba1]]
Former-commit-id: 5af551f6478926f3fa1bd9db68fc2b8c26afe64a
Former-commit-id: 5140aa92b59b29b7e6869bfb7210079965da1e07