prune.c: only print informational message in show_only or verbose mode
"git prune" reports removal of loose object files that are no longer
necessary only under the "-v" option, but unconditionally reports
removal of temporary files that are no longer needed.

The original thinking was that the presence of a leftover temporary
file should be an unusual occurrence that may indicate an earlier
failure of some sort, and the user may want to be reminded of it.
Removing an unnecessary loose object file, on the other hand, is
just part of the normal operation.  That is why the former is always
printed out and the latter only when -v is used.

But neither report is particularly useful.  Hide both of these
behind the "-v" option for consistency.

Signed-off-by: Brandon Casey <drafnel@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0ba6eb6c7fd59ca054918f7c0bbb17c5732ecdf6 [formerly 3e41fef7b8ce16e28a3fb1bfcc561867e5cdea16] [formerly 3e41fef7b8ce16e28a3fb1bfcc561867e5cdea16 [formerly 90b29cb7a8213a947e08c41ea4680ced7ac7c67c]]
Former-commit-id: 7134d4579ab7333da2e812b88301d95a6cfeded9
Former-commit-id: bda692a0510271a804adb76bad38b00bb6699885