check-docs: update non-command documentation list
The check-docs target looks at Documentation/git*txt and
complains if any entry does not have a matching command.
Therefore we need to explicitly ignore any entries which are
not meant to describe a command (like gitattributes.txt).
This list has grown stale over time, so let's bring it up to
date.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: bf305bd0561116241c1969122e1eec3f1cb9ef66 [formerly 8934eb5a687fd9c0989725f709a62e3f726c5941] [formerly 8934eb5a687fd9c0989725f709a62e3f726c5941 [formerly 1b7b522c2698ced2a60d5c26bf5b16e01bd45a75]]
Former-commit-id: 77ce94129dfc0d2bd3ac6ff29a6108987e0e61bc
Former-commit-id: c2a99090517afbe5f3d0eb4f4cb2c6dbcc58e429