command-list: mention git-credential-* helpers
These commands were never added to the command-list. Adding
them makes "make check-docs" run without complaint.
While we're at it, let's capitalize the first letter of
their one-line summaries to match the rest of the git
manpages.

The credential-cache--daemon command is somewhat special. It
is already ignored by check-docs because it contains a "--",
marking it as a non-interesting implementation detail. It
is, in fact, documented, but since the documentation
basically just redirects you to a more appropriate command
anyway, let's explicitly omit it so it is not mentioned in
git(1).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1dbcf1d8740dd0d76d8d424a4569d499a8a75c2e [formerly 1dbcf1d8740dd0d76d8d424a4569d499a8a75c2e [formerly c6632eba5a14791b1fa376cd59eb334c29732327]]
Former-commit-id: c08d7a82fa675082e5a52afd0ab0bb4d5ac7f637
Former-commit-id: c7ba5fc8b348b2a76efb97b5b5c419288b5d74f6