check-docs: list git-gui as a command
git-gui is already documented and mentioned in command-list,
but adding it to the Makefile makes sure it is so. We also
add its alias git-citool (which is also documented).

As a result, we can drop them from the special case
statement that avoids them being listed as "documented but
does not exist".

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2c79085bff93270356f71d3a5e7cd16d4801c8f7 [formerly 1e39a7b6370804d231d65e40be29b0dad43769d3] [formerly 1e39a7b6370804d231d65e40be29b0dad43769d3 [formerly eb28501e15c069d61c7dd7601fad2655449ebb14]]
Former-commit-id: 06577aaa4d1f3ead5337cc12f347762cd2137185
Former-commit-id: 19b9cffca5aa9f66247c64b4e461cd951113e887