rebase -i: use full onto sha1 in reflog
'git rebase' uses the full onto sha1 for the reflog message whereas 'git
rebase -i' uses the short sha1. This is not only inconsistent, but can
lead to problems when the reflog is inspected at a later time at which
that abbreviation may have become ambiguous.

Make 'rebase -i' use the full onto sha1, as well.

Signed-off-by: Michael J Gruber <git@drmicha.warpmail.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d4545f4b07bb9f582c802521a10c039ef929c4f2 [formerly 661b0154bc1675eb89d5de30ec55d2a2dd3431e2] [formerly 661b0154bc1675eb89d5de30ec55d2a2dd3431e2 [formerly 1af221ef5cb84c792303a0d16ca91197e2b70279]]
Former-commit-id: a82eb7656a614749df87e9b473bee5b6bf8c3c49
Former-commit-id: e69eb8ca2b04eae1c49afb87c08c572619626d72