do not send client agent unless server does first
Commit ff5effdf taught both clients and servers of the git protocol
to send an "agent" capability that just advertises their version for
statistics and debugging purposes.  The protocol-capabilities.txt
document however indicates that the client's advertisement is
actually a response, and should never include capabilities not
mentioned in the server's advertisement.

Adding the unconditional advertisement in the server programs was
OK, then, but the clients broke the protocol.  The server
implementation of git-core itself does not care, but at least one
does: the Google Code git server (or any server using Dulwich), will
hang up with an internal error upon seeing an unknown capability.

Instead, each client must record whether we saw an agent string from
the server, and respond with its agent only if the server mentioned
it first.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 456c29c4b1b25a305f518035699144b288093469 [formerly 456c29c4b1b25a305f518035699144b288093469 [formerly d50c3871632451b0cac066eea178198ac8a9dff1]]
Former-commit-id: 5240edf2d35ad2f436ee9bbbb99362ffe27bbafd
Former-commit-id: b620a1637f09dcdf776442e97d48ca62353a7127