git p4: make branch detection work with --use-client-spec
The bug report in http://stackoverflow.com/questions/11893688
observes that files are mapped into the wrong locations in
git when both --use-client-spec and --branch-detection are enabled.

Fix this by changing the relative path prefix to match discovered
branches when using a client spec.

The problem was likely introduced with ecb7cf9 (git-p4: rewrite view
handling, 2012-01-02).

Signed-off-by: Pete Wyckoff <pw@padd.com>
Tested-by: Matthew Korich <matthew@korich.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 20413b842b9a91fd35ee9d7dbd4e321cae47923c [formerly 5a3981875582effc223bb5ab4bc4fbee5a250b7f] [formerly 5a3981875582effc223bb5ab4bc4fbee5a250b7f [formerly 21ef5df4315b8f4f45bffb6af54d1ea338cb2cf5]]
Former-commit-id: 6d4155e567e182914caed8b7e58c0ac2a6a3f897
Former-commit-id: c977c2f0a000eaef2bd1c8b46cc496a98eb29612