man: git pull -r is a short for --rebase
Letting the "--rebase" option squat on the short-and-sweet single
letter option "-r" was an unintended accident and was not even
documented, but the short option seems to be already used in the
wild. Let's document it so that other options that begin with "r"
would not be tempted to steal it.

Signed-off-by: Miklos Vajna <vmiklos@suse.cz>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 764c8e94151c874825952d9bfd8f0e81de64af58 [formerly 764c8e94151c874825952d9bfd8f0e81de64af58 [formerly d9aa3610437e8cfab86fe3d1e3433400de15773d]]
Former-commit-id: ee8b4bd589fe97c247a575afa23b7f2f9f952339
Former-commit-id: 7b9bb53342dfa82661338dd358388bbd708d97e9