Documentation: update the introductory section
The second paragraph in the git(1) description section were meant to
guide people who are not ready to dive into this page away from here.
Referring migrating CVS users to another page before they get
acquainted with Git was somewhat out of place.  Move the reference to
the "FURTHER DOCUMENTATION" section and push that section down.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5db897ce15a8644fa0ddd9fbcec5b68d945a2d02 [formerly 5db897ce15a8644fa0ddd9fbcec5b68d945a2d02 [formerly 7687ae98e0b7e9ae12b1ef1c0c3cce0b609a633f]]
Former-commit-id: 930a403d6a8be0688c33785d7ce8c535951f7cef
Former-commit-id: 20602f71b2ad1c1f5e8a353a2b145724523207e8