git-svn.perl: consider all ranges for a given merge, instead of only tip-by-tip
Consider the case where you have trunk, branch1 of trunk, and branch2 of
branch1.  trunk is merged back into branch2, and then branch2 is
reintegrated into trunk.  The merge of branch2 into trunk will have
svn:mergeinfo property references to both branch1 and branch2.  When
git-svn fetches the commit that merges branch2 (check_cherry_pick),
it is necessary to eliminate the merged contents of branch1 as well as
branch2, or else the merge will be incorrectly ignored as a cherry-pick.

Signed-off-by: Steven Walter <stevenrwalter@gmail.com>
Reviewed-by: Sam Vilain <sam@vilain.net>
Signed-off-by: Eric Wong <normalperson@yhbt.net>

Former-commit-id: 18ea79c6ea5b7b3b7ef119b6edc020d89b9d1775 [formerly 81700ed4d0c189bf085c1878928b0a1dfee97473] [formerly 81700ed4d0c189bf085c1878928b0a1dfee97473 [formerly f271fad2661d2f26654e9b4b3ecb17ad0ea711e8]]
Former-commit-id: 314420b021ae5029603070360e7f3c62ecdc63f2
Former-commit-id: 7bfefb6b53114e405099a55d54a7f317f9e04ec1